Get countries

Get all country codes (alpha-3) based on the ISO 3166-1 standard, but adjusted for data from Statistics Norway.

Usage

get_countries(
  year = format(Sys.Date(), "%Y"),
  include_notes = FALSE,
  simplify = TRUE,
  raw_response = FALSE
)

Arguments

year

The year for which the codes should be valid for.

include_notes

If notes should be included or not.

simplify

If TRUE (default), a single data.frame is returned. Ignored if raw_response is set to TRUE.

raw_response

If TRUE a list of class noAPI is returned, including the raw httr2_response.

Value

data.frame or list

Details

Fetches all countries and country codes (alpha-3) for a given year based on the ISO 3166-1 standard. The list includes special codes XUK and XXX for unknown citizenship and stateless persons, respectively. These codes are used in official statistics from Statistics Norway.

All years from 1974 until present are supported. The function defaults to the current year.

If notes are enabled with include_notes, a column note will be added to the data.frame. Notes are stated with a code reference that has the following order (alpha-2, alpha-3, num-3, SSB-3), for example Norway (NO, NOR, 578, 000). See Statistics Norway's webpage for more information on the code specification.

The function returns a data.frame by default. If you prefer the output as a list you can set simplify to FALSE. This can be useful to keep programmatically track of failed queries. If you set raw_response to TRUE, the raw response from the API will be returned together with the parsed response. Note that the response will then be returned silently.

Examples

# Get country for current year
get_countries()
#>     year code                                        name
#> 1   2024  ABW                                       Aruba
#> 2   2024  AFG                                 Afghanistan
#> 3   2024  AGO                                      Angola
#> 4   2024  AIA                                    Anguilla
#> 5   2024  ALA                                       Åland
#> 6   2024  ALB                                     Albania
#> 7   2024  AND                                     Andorra
#> 8   2024  ARE                De forente arabiske emirater
#> 9   2024  ARG                                   Argentina
#> 10  2024  ARM                                     Armenia
#> 11  2024  ASM                            Amerikansk Samoa
#> 12  2024  ATA                                   Antarktis
#> 13  2024  ATF                   De franske sørterritorier
#> 14  2024  ATG                          Antigua og Barbuda
#> 15  2024  AUS                                   Australia
#> 16  2024  AUT                                   Østerrike
#> 17  2024  AZE                                Aserbajdsjan
#> 18  2024  BDI                                     Burundi
#> 19  2024  BEL                                      Belgia
#> 20  2024  BEN                                       Benin
#> 21  2024  BES             Bonaire, Sint Eustatius og Saba
#> 22  2024  BFA                                Burkina Faso
#> 23  2024  BGD                                  Bangladesh
#> 24  2024  BGR                                    Bulgaria
#> 25  2024  BHR                                     Bahrain
#> 26  2024  BHS                                     Bahamas
#> 27  2024  BIH                          Bosnia-Hercegovina
#> 28  2024  BLM                            Saint-Barthélemy
#> 29  2024  BLR                                     Belarus
#> 30  2024  BLZ                                      Belize
#> 31  2024  BMU                                     Bermuda
#> 32  2024  BOL                                     Bolivia
#> 33  2024  BRA                                      Brasil
#> 34  2024  BRB                                    Barbados
#> 35  2024  BRN                                      Brunei
#> 36  2024  BTN                                      Bhutan
#> 37  2024  BVT                                   Bouvetøya
#> 38  2024  BWA                                    Botswana
#> 39  2024  CAF             Den sentralafrikanske republikk
#> 40  2024  CAN                                      Canada
#> 41  2024  CCK                                  Kokosøyene
#> 42  2024  CHE                                      Sveits
#> 43  2024  CHL                                       Chile
#> 44  2024  CHN                                        Kina
#> 45  2024  CIV                            Elfenbeinskysten
#> 46  2024  CMR                                     Kamerun
#> 47  2024  COD                                       Kongo
#> 48  2024  COG                         Kongo - Brazzaville
#> 49  2024  COK                                   Cookøyene
#> 50  2024  COL                                    Colombia
#> 51  2024  COM                                    Komorene
#> 52  2024  CPV                                  Kapp Verde
#> 53  2024  CRI                                  Costa Rica
#> 54  2024  CUB                                        Cuba
#> 55  2024  CUW                                     Curaçao
#> 56  2024  CXR                                Christmasøya
#> 57  2024  CYM                                 Caymanøyene
#> 58  2024  CYP                                      Kypros
#> 59  2024  CZE                                    Tsjekkia
#> 60  2024  DEU                                    Tyskland
#> 61  2024  DJI                                    Djibouti
#> 62  2024  DMA                                    Dominica
#> 63  2024  DNK                                     Danmark
#> 64  2024  DOM                  Den dominikanske republikk
#> 65  2024  DZA                                     Algerie
#> 66  2024  ECU                                     Ecuador
#> 67  2024  EGY                                       Egypt
#> 68  2024  ERI                                     Eritrea
#> 69  2024  ESH                                 Vest-Sahara
#> 70  2024  ESP                                      Spania
#> 71  2024  EST                                     Estland
#> 72  2024  ETH                                     Etiopia
#> 73  2024  FIN                                     Finland
#> 74  2024  FJI                                        Fiji
#> 75  2024  FLK                              Falklandsøyene
#> 76  2024  FRA                                   Frankrike
#> 77  2024  FRO                                    Færøyene
#> 78  2024  FSM                      Mikronesiaføderasjonen
#> 79  2024  GAB                                       Gabon
#> 80  2024  GBR                               Storbritannia
#> 81  2024  GEO                                     Georgia
#> 82  2024  GGY                                    Guernsey
#> 83  2024  GHA                                       Ghana
#> 84  2024  GIB                                   Gibraltar
#> 85  2024  GIN                                      Guinea
#> 86  2024  GLP                                  Guadeloupe
#> 87  2024  GMB                                      Gambia
#> 88  2024  GNB                               Guinea-Bissau
#> 89  2024  GNQ                           Ekvatorial-Guinea
#> 90  2024  GRC                                      Hellas
#> 91  2024  GRD                                     Grenada
#> 92  2024  GRL                                    Grønland
#> 93  2024  GTM                                   Guatemala
#> 94  2024  GUF                               Fransk Guyana
#> 95  2024  GUM                                        Guam
#> 96  2024  GUY                                      Guyana
#> 97  2024  HKG                                    Hongkong
#> 98  2024  HMD                     Heard- og McDonaldøyene
#> 99  2024  HND                                    Honduras
#> 100 2024  HRV                                     Kroatia
#> 101 2024  HTI                                       Haiti
#> 102 2024  HUN                                      Ungarn
#> 103 2024  IDN                                   Indonesia
#> 104 2024  IMN                                         Man
#> 105 2024  IND                                       India
#> 106 2024  IOT       Det britiske territorium i Indiahavet
#> 107 2024  IRL                                      Irland
#> 108 2024  IRN                                        Iran
#> 109 2024  IRQ                                        Irak
#> 110 2024  ISL                                      Island
#> 111 2024  ISR                                      Israel
#> 112 2024  ITA                                      Italia
#> 113 2024  JAM                                     Jamaica
#> 114 2024  JEY                                      Jersey
#> 115 2024  JOR                                      Jordan
#> 116 2024  JPN                                       Japan
#> 117 2024  KAZ                                  Kasakhstan
#> 118 2024  KEN                                       Kenya
#> 119 2024  KGZ                                 Kirgisistan
#> 120 2024  KHM                                   Kambodsja
#> 121 2024  KIR                                    Kiribati
#> 122 2024  KNA                          St. Kitts og Nevis
#> 123 2024  KOR                                   Sør-Korea
#> 124 2024  KWT                                      Kuwait
#> 125 2024  LAO                                        Laos
#> 126 2024  LBN                                     Libanon
#> 127 2024  LBR                                     Liberia
#> 128 2024  LBY                                       Libya
#> 129 2024  LCA                                   St. Lucia
#> 130 2024  LIE                               Liechtenstein
#> 131 2024  LKA                                   Sri Lanka
#> 132 2024  LSO                                     Lesotho
#> 133 2024  LTU                                     Litauen
#> 134 2024  LUX                                  Luxembourg
#> 135 2024  LVA                                      Latvia
#> 136 2024  MAC                                       Macao
#> 137 2024  MAF                           Saint-Martin (FR)
#> 138 2024  MAR                                     Marokko
#> 139 2024  MCO                                      Monaco
#> 140 2024  MDA                                     Moldova
#> 141 2024  MDG                                  Madagaskar
#> 142 2024  MDV                                   Maldivene
#> 143 2024  MEX                                      Mexico
#> 144 2024  MHL                               Marshalløyene
#> 145 2024  MKD                              Nord-Makedonia
#> 146 2024  MLI                                        Mali
#> 147 2024  MLT                                       Malta
#> 148 2024  MMR                                     Myanmar
#> 149 2024  MNE                                  Montenegro
#> 150 2024  MNG                                    Mongolia
#> 151 2024  MNP                              Nord-Marianene
#> 152 2024  MOZ                                    Mosambik
#> 153 2024  MRT                                  Mauritania
#> 154 2024  MSR                                  Montserrat
#> 155 2024  MTQ                                  Martinique
#> 156 2024  MUS                                   Mauritius
#> 157 2024  MWI                                      Malawi
#> 158 2024  MYS                                    Malaysia
#> 159 2024  MYT                                     Mayotte
#> 160 2024  NAM                                     Namibia
#> 161 2024  NCL                                Ny-Caledonia
#> 162 2024  NER                                       Niger
#> 163 2024  NFK                                  Norfolkøya
#> 164 2024  NGA                                     Nigeria
#> 165 2024  NIC                                   Nicaragua
#> 166 2024  NIU                                        Niue
#> 167 2024  NLD                                   Nederland
#> 168 2024  NOR                                       Norge
#> 169 2024  NPL                                       Nepal
#> 170 2024  NRU                                       Nauru
#> 171 2024  NZL                                 New Zealand
#> 172 2024  OMN                                        Oman
#> 173 2024  PAK                                    Pakistan
#> 174 2024  PAN                                      Panama
#> 175 2024  PCN                                    Pitcairn
#> 176 2024  PER                                        Peru
#> 177 2024  PHL                                 Filippinene
#> 178 2024  PLW                                       Palau
#> 179 2024  PNG                             Papua Ny-Guinea
#> 180 2024  POL                                       Polen
#> 181 2024  PRI                                 Puerto Rico
#> 182 2024  PRK                                  Nord-Korea
#> 183 2024  PRT                                    Portugal
#> 184 2024  PRY                                    Paraguay
#> 185 2024  PSE                                   Palestina
#> 186 2024  PYF                            Fransk Polynesia
#> 187 2024  QAT                                       Qatar
#> 188 2024  REU                                     Réunion
#> 189 2024  ROU                                     Romania
#> 190 2024  RUS                                    Russland
#> 191 2024  RWA                                      Rwanda
#> 192 2024  SAU                                Saudi-Arabia
#> 193 2024  SDN                                       Sudan
#> 194 2024  SEN                                     Senegal
#> 195 2024  SGP                                   Singapore
#> 196 2024  SGS            Sør-Georgia og Sør-Sandwichøyene
#> 197 2024  SHN Sankt Helena, Ascension og Tristan da Cunha
#> 198 2024  SJM                       Svalbard og Jan Mayen
#> 199 2024  SLB                                Salomonøyene
#> 200 2024  SLE                                Sierra Leone
#> 201 2024  SLV                                 El Salvador
#> 202 2024  SMR                                  San Marino
#> 203 2024  SOM                                     Somalia
#> 204 2024  SPM                    Saint-Pierre-et-Miquelon
#> 205 2024  SRB                                      Serbia
#> 206 2024  SSD                                   Sør-Sudan
#> 207 2024  STP                        São Tomé og Príncipe
#> 208 2024  SUR                                     Surinam
#> 209 2024  SVK                                    Slovakia
#> 210 2024  SVN                                    Slovenia
#> 211 2024  SWE                                     Sverige
#> 212 2024  SWZ                                    Eswatini
#> 213 2024  SXM                           Sint Maarten (NL)
#> 214 2024  SYC                                 Seychellene
#> 215 2024  SYR                                       Syria
#> 216 2024  TCA                       Turks- og Caicosøyene
#> 217 2024  TCD                                       Tsjad
#> 218 2024  TGO                                        Togo
#> 219 2024  THA                                    Thailand
#> 220 2024  TJK                                Tadsjikistan
#> 221 2024  TKL                                     Tokelau
#> 222 2024  TKM                                Turkmenistan
#> 223 2024  TLS                                   Øst-Timor
#> 224 2024  TON                                       Tonga
#> 225 2024  TTO                          Trinidad og Tobago
#> 226 2024  TUN                                     Tunisia
#> 227 2024  TUR                                      Tyrkia
#> 228 2024  TUV                                      Tuvalu
#> 229 2024  TWN                                      Taiwan
#> 230 2024  TZA                                    Tanzania
#> 231 2024  UGA                                      Uganda
#> 232 2024  UKR                                     Ukraina
#> 233 2024  UMI                           USAs ytre småøyer
#> 234 2024  URY                                     Uruguay
#> 235 2024  USA                                         USA
#> 236 2024  UZB                                  Usbekistan
#> 237 2024  VAT                               Vatikanstaten
#> 238 2024  VCT                  St. Vincent og Grenadinene
#> 239 2024  VEN                                   Venezuela
#> 240 2024  VGB                      De britiske Jomfruøyer
#> 241 2024  VIR                   De amerikanske Jomfruøyer
#> 242 2024  VNM                                     Vietnam
#> 243 2024  VUT                                     Vanuatu
#> 244 2024  WLF                      Wallis- og Futunaøyene
#> 245 2024  WSM                                       Samoa
#> 246 2024  XUK                                    Uoppgitt
#> 247 2024  XXK                                      Kosovo
#> 248 2024  XXX                                    Statsløs
#> 249 2024  YEM                                       Jemen
#> 250 2024  ZAF                                  Sør-Afrika
#> 251 2024  ZMB                                      Zambia
#> 252 2024  ZWE                                    Zimbabwe