Get countries

Get all country codes (alpha-3) based on the ISO 3166-1 standard, but adjusted for data from Statistics Norway.

Usage

get_countries(
  year = format(Sys.Date(), "%Y"),
  include_notes = FALSE,
  simplify = TRUE,
  raw_response = FALSE
)

Arguments

year

The year for which the codes should be valid for.

include_notes

If notes should be included or not.

simplify

If TRUE (default), a single data.frame is returned. Ignored if raw_response is set to TRUE.

raw_response

If TRUE a list of class noAPI is returned, including the raw httr2_response.

Value

data.frame or list

Details

Fetches all countries and country codes (alpha-3) for a given year based on the ISO 3166-1 standard. The list includes special codes XUK and XXX for unknown citizenship and stateless persons, respectively. These codes are used in official statistics from Statistics Norway.

All years from 1974 until present are supported. The function defaults to the current year.

If notes are enabled with include_notes, a column note will be added to the data.frame. Notes are stated with a code reference that has the following order (alpha-2, alpha-3, num-3, SSB-3), for example Norway (NO, NOR, 578, 000). See Statistics Norway's webpage for more information on the code specification.

The function returns a data.frame by default. If you prefer the output as a list you can set simplify to FALSE. This can be useful to keep programmatically track of failed queries. If you set raw_response to TRUE, the raw response from the API will be returned together with the parsed response. Note that the response will then be returned silently.

Examples

# Get country for current year
get_countries()
#>     year code                                        name
#> 1   2023  ABW                                       Aruba
#> 2   2023  AFG                                 Afghanistan
#> 3   2023  AGO                                      Angola
#> 4   2023  AIA                                    Anguilla
#> 5   2023  ALA                                       Åland
#> 6   2023  ALB                                     Albania
#> 7   2023  AND                                     Andorra
#> 8   2023  ARE                De forente arabiske emirater
#> 9   2023  ARG                                   Argentina
#> 10  2023  ARM                                     Armenia
#> 11  2023  ASM                            Amerikansk Samoa
#> 12  2023  ATA                                   Antarktis
#> 13  2023  ATF                   De franske sørterritorier
#> 14  2023  ATG                          Antigua og Barbuda
#> 15  2023  AUS                                   Australia
#> 16  2023  AUT                                   Østerrike
#> 17  2023  AZE                                Aserbajdsjan
#> 18  2023  BDI                                     Burundi
#> 19  2023  BEL                                      Belgia
#> 20  2023  BEN                                       Benin
#> 21  2023  BES             Bonaire, Sint Eustatius og Saba
#> 22  2023  BFA                                Burkina Faso
#> 23  2023  BGD                                  Bangladesh
#> 24  2023  BGR                                    Bulgaria
#> 25  2023  BHR                                     Bahrain
#> 26  2023  BHS                                     Bahamas
#> 27  2023  BIH                          Bosnia-Hercegovina
#> 28  2023  BLM                            Saint-Barthélemy
#> 29  2023  BLR                                     Belarus
#> 30  2023  BLZ                                      Belize
#> 31  2023  BMU                                     Bermuda
#> 32  2023  BOL                                     Bolivia
#> 33  2023  BRA                                      Brasil
#> 34  2023  BRB                                    Barbados
#> 35  2023  BRN                                      Brunei
#> 36  2023  BTN                                      Bhutan
#> 37  2023  BVT                                   Bouvetøya
#> 38  2023  BWA                                    Botswana
#> 39  2023  CAF             Den sentralafrikanske republikk
#> 40  2023  CAN                                      Canada
#> 41  2023  CCK                                  Kokosøyene
#> 42  2023  CHE                                      Sveits
#> 43  2023  CHL                                       Chile
#> 44  2023  CHN                                        Kina
#> 45  2023  CIV                            Elfenbeinskysten
#> 46  2023  CMR                                     Kamerun
#> 47  2023  COD                                       Kongo
#> 48  2023  COG                         Kongo - Brazzaville
#> 49  2023  COK                                   Cookøyene
#> 50  2023  COL                                    Colombia
#> 51  2023  COM                                    Komorene
#> 52  2023  CPV                                  Kapp Verde
#> 53  2023  CRI                                  Costa Rica
#> 54  2023  CUB                                        Cuba
#> 55  2023  CUW                                     Curaçao
#> 56  2023  CXR                                Christmasøya
#> 57  2023  CYM                                 Caymanøyene
#> 58  2023  CYP                                      Kypros
#> 59  2023  CZE                                    Tsjekkia
#> 60  2023  DEU                                    Tyskland
#> 61  2023  DJI                                    Djibouti
#> 62  2023  DMA                                    Dominica
#> 63  2023  DNK                                     Danmark
#> 64  2023  DOM                  Den dominikanske republikk
#> 65  2023  DZA                                     Algerie
#> 66  2023  ECU                                     Ecuador
#> 67  2023  EGY                                       Egypt
#> 68  2023  ERI                                     Eritrea
#> 69  2023  ESH                                 Vest-Sahara
#> 70  2023  ESP                                      Spania
#> 71  2023  EST                                     Estland
#> 72  2023  ETH                                     Etiopia
#> 73  2023  FIN                                     Finland
#> 74  2023  FJI                                        Fiji
#> 75  2023  FLK                              Falklandsøyene
#> 76  2023  FRA                                   Frankrike
#> 77  2023  FRO                                    Færøyene
#> 78  2023  FSM                      Mikronesiaføderasjonen
#> 79  2023  GAB                                       Gabon
#> 80  2023  GBR                               Storbritannia
#> 81  2023  GEO                                     Georgia
#> 82  2023  GGY                                    Guernsey
#> 83  2023  GHA                                       Ghana
#> 84  2023  GIB                                   Gibraltar
#> 85  2023  GIN                                      Guinea
#> 86  2023  GLP                                  Guadeloupe
#> 87  2023  GMB                                      Gambia
#> 88  2023  GNB                               Guinea-Bissau
#> 89  2023  GNQ                           Ekvatorial-Guinea
#> 90  2023  GRC                                      Hellas
#> 91  2023  GRD                                     Grenada
#> 92  2023  GRL                                    Grønland
#> 93  2023  GTM                                   Guatemala
#> 94  2023  GUF                               Fransk Guyana
#> 95  2023  GUM                                        Guam
#> 96  2023  GUY                                      Guyana
#> 97  2023  HKG                                    Hongkong
#> 98  2023  HMD                     Heard- og McDonaldøyene
#> 99  2023  HND                                    Honduras
#> 100 2023  HRV                                     Kroatia
#> 101 2023  HTI                                       Haiti
#> 102 2023  HUN                                      Ungarn
#> 103 2023  IDN                                   Indonesia
#> 104 2023  IMN                                         Man
#> 105 2023  IND                                       India
#> 106 2023  IOT       Det britiske territorium i Indiahavet
#> 107 2023  IRL                                      Irland
#> 108 2023  IRN                                        Iran
#> 109 2023  IRQ                                        Irak
#> 110 2023  ISL                                      Island
#> 111 2023  ISR                                      Israel
#> 112 2023  ITA                                      Italia
#> 113 2023  JAM                                     Jamaica
#> 114 2023  JEY                                      Jersey
#> 115 2023  JOR                                      Jordan
#> 116 2023  JPN                                       Japan
#> 117 2023  KAZ                                  Kasakhstan
#> 118 2023  KEN                                       Kenya
#> 119 2023  KGZ                                 Kirgisistan
#> 120 2023  KHM                                   Kambodsja
#> 121 2023  KIR                                    Kiribati
#> 122 2023  KNA                          St. Kitts og Nevis
#> 123 2023  KOR                                   Sør-Korea
#> 124 2023  KWT                                      Kuwait
#> 125 2023  LAO                                        Laos
#> 126 2023  LBN                                     Libanon
#> 127 2023  LBR                                     Liberia
#> 128 2023  LBY                                       Libya
#> 129 2023  LCA                                   St. Lucia
#> 130 2023  LIE                               Liechtenstein
#> 131 2023  LKA                                   Sri Lanka
#> 132 2023  LSO                                     Lesotho
#> 133 2023  LTU                                     Litauen
#> 134 2023  LUX                                  Luxembourg
#> 135 2023  LVA                                      Latvia
#> 136 2023  MAC                                       Macao
#> 137 2023  MAF                           Saint-Martin (FR)
#> 138 2023  MAR                                     Marokko
#> 139 2023  MCO                                      Monaco
#> 140 2023  MDA                                     Moldova
#> 141 2023  MDG                                  Madagaskar
#> 142 2023  MDV                                   Maldivene
#> 143 2023  MEX                                      Mexico
#> 144 2023  MHL                               Marshalløyene
#> 145 2023  MKD                              Nord-Makedonia
#> 146 2023  MLI                                        Mali
#> 147 2023  MLT                                       Malta
#> 148 2023  MMR                                     Myanmar
#> 149 2023  MNE                                  Montenegro
#> 150 2023  MNG                                    Mongolia
#> 151 2023  MNP                              Nord-Marianene
#> 152 2023  MOZ                                    Mosambik
#> 153 2023  MRT                                  Mauritania
#> 154 2023  MSR                                  Montserrat
#> 155 2023  MTQ                                  Martinique
#> 156 2023  MUS                                   Mauritius
#> 157 2023  MWI                                      Malawi
#> 158 2023  MYS                                    Malaysia
#> 159 2023  MYT                                     Mayotte
#> 160 2023  NAM                                     Namibia
#> 161 2023  NCL                                Ny-Caledonia
#> 162 2023  NER                                       Niger
#> 163 2023  NFK                                  Norfolkøya
#> 164 2023  NGA                                     Nigeria
#> 165 2023  NIC                                   Nicaragua
#> 166 2023  NIU                                        Niue
#> 167 2023  NLD                                   Nederland
#> 168 2023  NOR                                       Norge
#> 169 2023  NPL                                       Nepal
#> 170 2023  NRU                                       Nauru
#> 171 2023  NZL                                 New Zealand
#> 172 2023  OMN                                        Oman
#> 173 2023  PAK                                    Pakistan
#> 174 2023  PAN                                      Panama
#> 175 2023  PCN                                    Pitcairn
#> 176 2023  PER                                        Peru
#> 177 2023  PHL                                 Filippinene
#> 178 2023  PLW                                       Palau
#> 179 2023  PNG                             Papua Ny-Guinea
#> 180 2023  POL                                       Polen
#> 181 2023  PRI                                 Puerto Rico
#> 182 2023  PRK                                  Nord-Korea
#> 183 2023  PRT                                    Portugal
#> 184 2023  PRY                                    Paraguay
#> 185 2023  PSE                                   Palestina
#> 186 2023  PYF                            Fransk Polynesia
#> 187 2023  QAT                                       Qatar
#> 188 2023  REU                                     Réunion
#> 189 2023  ROU                                     Romania
#> 190 2023  RUS                                    Russland
#> 191 2023  RWA                                      Rwanda
#> 192 2023  SAU                                Saudi-Arabia
#> 193 2023  SDN                                       Sudan
#> 194 2023  SEN                                     Senegal
#> 195 2023  SGP                                   Singapore
#> 196 2023  SGS            Sør-Georgia og Sør-Sandwichøyene
#> 197 2023  SHN Sankt Helena, Ascension og Tristan da Cunha
#> 198 2023  SJM                       Svalbard og Jan Mayen
#> 199 2023  SLB                                Salomonøyene
#> 200 2023  SLE                                Sierra Leone
#> 201 2023  SLV                                 El Salvador
#> 202 2023  SMR                                  San Marino
#> 203 2023  SOM                                     Somalia
#> 204 2023  SPM                    Saint-Pierre-et-Miquelon
#> 205 2023  SRB                                      Serbia
#> 206 2023  SSD                                   Sør-Sudan
#> 207 2023  STP                        São Tomé og Príncipe
#> 208 2023  SUR                                     Surinam
#> 209 2023  SVK                                    Slovakia
#> 210 2023  SVN                                    Slovenia
#> 211 2023  SWE                                     Sverige
#> 212 2023  SWZ                                    Eswatini
#> 213 2023  SXM                           Sint Maarten (NL)
#> 214 2023  SYC                                 Seychellene
#> 215 2023  SYR                                       Syria
#> 216 2023  TCA                       Turks- og Caicosøyene
#> 217 2023  TCD                                       Tsjad
#> 218 2023  TGO                                        Togo
#> 219 2023  THA                                    Thailand
#> 220 2023  TJK                                Tadsjikistan
#> 221 2023  TKL                                     Tokelau
#> 222 2023  TKM                                Turkmenistan
#> 223 2023  TLS                                   Øst-Timor
#> 224 2023  TON                                       Tonga
#> 225 2023  TTO                          Trinidad og Tobago
#> 226 2023  TUN                                     Tunisia
#> 227 2023  TUR                                      Tyrkia
#> 228 2023  TUV                                      Tuvalu
#> 229 2023  TWN                                      Taiwan
#> 230 2023  TZA                                    Tanzania
#> 231 2023  UGA                                      Uganda
#> 232 2023  UKR                                     Ukraina
#> 233 2023  UMI                           USAs ytre småøyer
#> 234 2023  URY                                     Uruguay
#> 235 2023  USA                                         USA
#> 236 2023  UZB                                  Usbekistan
#> 237 2023  VAT                               Vatikanstaten
#> 238 2023  VCT                  St. Vincent og Grenadinene
#> 239 2023  VEN                                   Venezuela
#> 240 2023  VGB                      De britiske Jomfruøyer
#> 241 2023  VIR                   De amerikanske Jomfruøyer
#> 242 2023  VNM                                     Vietnam
#> 243 2023  VUT                                     Vanuatu
#> 244 2023  WLF                      Wallis- og Futunaøyene
#> 245 2023  WSM                                       Samoa
#> 246 2023  XUK                                    Uoppgitt
#> 247 2023  XXK                                      Kosovo
#> 248 2023  XXX                                    Statsløs
#> 249 2023  YEM                                       Jemen
#> 250 2023  ZAF                                  Sør-Afrika
#> 251 2023  ZMB                                      Zambia
#> 252 2023  ZWE                                    Zimbabwe