Get exchange rate

Fetch exchange rates from Norges Bank.

Usage

get_exchange_rate(
  currency,
  frequency = c("daily", "monthly", "annual"),
  n_obs = 10,
  start = NULL,
  end = NULL,
  simplify = TRUE,
  raw_response = FALSE
)

Arguments

currency

The currency to fetch data for.

frequency

Frequency of data. Can be daily, monthly or annual. Defaults to daily.

n_obs

The number of observations to fetch. See details.

start

The first date to fetch data for.

end

The last date to fetch data for.

simplify

If TRUE (default), a single data.frame is returned. Ignored if raw_response is set to TRUE.

raw_response

If TRUE a list of class noAPI is returned, including the raw httr2_response.

Value

data.frame or list

Details

Currency must be formatted as a three letter ISO 3166 code. A list of available country codes can be found at Norges Bank. You can fetch observations for several currencies by passing a character vector. If you want to include all currencies, you can set the currency argument to ALL. Note that the number of observations is limited to 20 when retrieving all currencies.

By default, the 10 latest observations are returned. You can adjust the number of observations by setting n_obs to a positive integer. You can also fetch exchange rates for a specific time period by setting the arguments start and end. Both arguments accept a date in the format %Y-%m-%d (YYYY-MM-DD).

The function returns a data.frame by default. If you prefer the output as a list you can set simplify to FALSE. This can be useful to keep programmatically track of failed queries. If you set raw_response to TRUE, the raw response from the API will be returned together with the parsed response. Note that the response will then be returned silently.

Examples

# Get latest daily exchange rates
get_exchange_rate('EUR')
#>    currency frequency exch_rate decimals calculated unit_multiplier  date_from
#> 1       EUR  Business   11.5905        4      false           Units 2024-04-08
#> 2       EUR  Business   11.5930        4      false           Units 2024-04-09
#> 3       EUR  Business   11.5665        4      false           Units 2024-04-10
#> 4       EUR  Business   11.6180        4      false           Units 2024-04-11
#> 5       EUR  Business   11.5870        4      false           Units 2024-04-12
#> 6       EUR  Business   11.6260        4      false           Units 2024-04-15
#> 7       EUR  Business   11.6745        4      false           Units 2024-04-16
#> 8       EUR  Business   11.6960        4      false           Units 2024-04-17
#> 9       EUR  Business   11.7375        4      false           Units 2024-04-18
#> 10      EUR  Business   11.7650        4      false           Units 2024-04-19
#>       date_to       date
#> 1  2024-04-08 2024-04-08
#> 2  2024-04-09 2024-04-09
#> 3  2024-04-10 2024-04-10
#> 4  2024-04-11 2024-04-11
#> 5  2024-04-12 2024-04-12
#> 6  2024-04-15 2024-04-15
#> 7  2024-04-16 2024-04-16
#> 8  2024-04-17 2024-04-17
#> 9  2024-04-18 2024-04-18
#> 10 2024-04-19 2024-04-19
# Get monthly exchange rates for multiple currencies
get_exchange_rate(c('EUR', 'USD'), frequency = 'monthly')
#>    currency frequency exch_rate decimals calculated unit_multiplier  date_from
#> 1       EUR   Monthly   11.7164        4      false           Units 2023-06-01
#> 2       EUR   Monthly   11.3474        4      false           Units 2023-07-01
#> 3       EUR   Monthly   11.4127        4      false           Units 2023-08-01
#> 4       EUR   Monthly   11.4525        4      false           Units 2023-09-01
#> 5       EUR   Monthly   11.6284        4      false           Units 2023-10-01
#> 6       EUR   Monthly   11.7958        4      false           Units 2023-11-01
#> 7       EUR   Monthly   11.5333        4      false           Units 2023-12-01
#> 8       EUR   Monthly   11.3501        4      false           Units 2024-01-01
#> 9       EUR   Monthly   11.3843        4      false           Units 2024-02-01
#> 10      EUR   Monthly   11.5120        4      false           Units 2024-03-01
#> 11      USD   Monthly   10.8099        4      false           Units 2023-06-01
#> 12      USD   Monthly   10.2644        4      false           Units 2023-07-01
#> 13      USD   Monthly   10.4629        4      false           Units 2023-08-01
#> 14      USD   Monthly   10.7197        4      false           Units 2023-09-01
#> 15      USD   Monthly   11.0089        4      false           Units 2023-10-01
#> 16      USD   Monthly   10.9161        4      false           Units 2023-11-01
#> 17      USD   Monthly   10.5810        4      false           Units 2023-12-01
#> 18      USD   Monthly   10.4083        4      false           Units 2024-01-01
#> 19      USD   Monthly   10.5463        4      false           Units 2024-02-01
#> 20      USD   Monthly   10.5857        4      false           Units 2024-03-01
#>       date_to    date
#> 1  2023-06-30 2023-06
#> 2  2023-07-31 2023-07
#> 3  2023-08-31 2023-08
#> 4  2023-09-30 2023-09
#> 5  2023-10-31 2023-10
#> 6  2023-11-30 2023-11
#> 7  2023-12-31 2023-12
#> 8  2024-01-31 2024-01
#> 9  2024-02-29 2024-02
#> 10 2024-03-31 2024-03
#> 11 2023-06-30 2023-06
#> 12 2023-07-31 2023-07
#> 13 2023-08-31 2023-08
#> 14 2023-09-30 2023-09
#> 15 2023-10-31 2023-10
#> 16 2023-11-30 2023-11
#> 17 2023-12-31 2023-12
#> 18 2024-01-31 2024-01
#> 19 2024-02-29 2024-02
#> 20 2024-03-31 2024-03
# Get last 5 observations for all currencies
get_exchange_rate('ALL', n_obs = 5)
#>     currency frequency   exch_rate decimals calculated unit_multiplier
#> 1        HUF  Business    2.948900        4      false        Hundreds
#> 2        HUF  Business    2.958300        4      false        Hundreds
#> 3        HUF  Business    2.976100        4      false        Hundreds
#> 4        HUF  Business    2.979600        4      false        Hundreds
#> 5        HUF  Business    2.976200        4      false        Hundreds
#> 6        DKK  Business  155.830000        2      false        Hundreds
#> 7        DKK  Business  156.480000        2      false        Hundreds
#> 8        DKK  Business  156.760000        2      false        Hundreds
#> 9        DKK  Business  157.310000        2      false        Hundreds
#> 10       DKK  Business  157.660000        2      false        Hundreds
#> 11       RUB  Business   11.040000        3      false        Hundreds
#> 12       RUB  Business   10.539000        3      false        Hundreds
#> 13       RUB  Business   10.777000        3      false        Hundreds
#> 14       RUB  Business    8.613000        3      false        Hundreds
#> 15       RUB  Business    8.413000        3      false        Hundreds
#> 16       MMK  Business    0.519500        4      false        Hundreds
#> 17       MMK  Business    0.522600        4      false        Hundreds
#> 18       MMK  Business    0.523500        4      false        Hundreds
#> 19       MMK  Business    0.523400        4      false        Hundreds
#> 20       MMK  Business    0.525900        4      false        Hundreds
#> 21       XDR  Business   14.364350        5       true           Units
#> 22       XDR  Business   14.435290        5       true           Units
#> 23       XDR  Business   14.461560        5       true           Units
#> 24       XDR  Business   14.477550        5       true           Units
#> 25       XDR  Business   14.531290        5       true           Units
#> 26       TRY  Business   33.670000        2      false        Hundreds
#> 27       TRY  Business   33.770000        2      false        Hundreds
#> 28       TRY  Business   33.800000        2      false        Hundreds
#> 29       TRY  Business   33.820000        2      false        Hundreds
#> 30       TRY  Business   33.880000        2      false        Hundreds
#> 31       RON  Business  233.730000        2      false        Hundreds
#> 32       RON  Business  234.600000        2      false        Hundreds
#> 33       RON  Business  235.060000        2      false        Hundreds
#> 34       RON  Business  235.890000        2      false        Hundreds
#> 35       RON  Business  236.420000        2      false        Hundreds
#> 36       NZD  Business    6.479400        4      false           Units
#> 37       NZD  Business    6.460000        4      false           Units
#> 38       NZD  Business    6.493400        4      false           Units
#> 39       NZD  Business    6.516100        4      false           Units
#> 40       NZD  Business    6.498200        4      false           Units
#> 41       JPY  Business    7.086900        4      false        Hundreds
#> 42       JPY  Business    7.095200        4      false        Hundreds
#> 43       JPY  Business    7.108300        4      false        Hundreds
#> 44       JPY  Business    7.121400        4      false        Hundreds
#> 45       JPY  Business    7.144200        4      false        Hundreds
#> 46       HKD  Business    1.393700        4      false           Units
#> 47       HKD  Business    1.401400        4      false           Units
#> 48       HKD  Business    1.404000        4      false           Units
#> 49       HKD  Business    1.403400        4      false           Units
#> 50       HKD  Business    1.409900        4      false           Units
#> 51       CZK  Business   45.909000        3      false        Hundreds
#> 52       CZK  Business   46.309000        3      false        Hundreds
#> 53       CZK  Business   46.361000        3      false        Hundreds
#> 54       CZK  Business   46.478000        3      false        Hundreds
#> 55       CZK  Business   46.563000        3      false        Hundreds
#> 56       CAD  Business    7.938500        4      false           Units
#> 57       CAD  Business    7.959200        4      false           Units
#> 58       CAD  Business    7.961300        4      false           Units
#> 59       CAD  Business    7.991800        4      false           Units
#> 60       CAD  Business    8.022000        4      false           Units
#> 61       BDT  Business    9.940000        2      false        Hundreds
#> 62       BDT  Business   10.000000        2      false        Hundreds
#> 63       BDT  Business   10.020000        2      false        Hundreds
#> 64       BDT  Business   10.010000        2      false        Hundreds
#> 65       BDT  Business   10.060000        2      false        Hundreds
#> 66       USD  Business   10.910300        4      false           Units
#> 67       USD  Business   10.975400        4      false           Units
#> 68       USD  Business   10.994500        4      false           Units
#> 69       USD  Business   10.991200        4      false           Units
#> 70       USD  Business   11.043800        4      false           Units
#> 71       THB  Business   29.716000        3      false        Hundreds
#> 72       THB  Business   29.967000        3      false        Hundreds
#> 73       THB  Business   29.852000        3      false        Hundreds
#> 74       THB  Business   29.904000        3      false        Hundreds
#> 75       THB  Business   29.946000        3      false        Hundreds
#> 76       PLN  Business    2.707600        4      false           Units
#> 77       PLN  Business    2.687800        4      false           Units
#> 78       PLN  Business    2.688200        4      false           Units
#> 79       PLN  Business    2.713600        4      false           Units
#> 80       PLN  Business    2.717100        4      false           Units
#> 81       AUD  Business    7.071300        4      false           Units
#> 82       AUD  Business    7.042200        4      false           Units
#> 83       AUD  Business    7.064500        4      false           Units
#> 84       AUD  Business    7.084900        4      false           Units
#> 85       AUD  Business    7.078000        4      false           Units
#> 86       MYR  Business    2.283000        4      false           Units
#> 87       MYR  Business    2.290300        4      false           Units
#> 88       MYR  Business    2.294600        4      false           Units
#> 89       MYR  Business    2.297700        4      false           Units
#> 90       MYR  Business    2.308700        4      false           Units
#> 91       INR  Business   13.078000        3      false        Hundreds
#> 92       INR  Business   13.133000        3      false        Hundreds
#> 93       INR  Business   13.140000        3      false        Hundreds
#> 94       INR  Business   13.160000        3      false        Hundreds
#> 95       INR  Business   13.231000        3      false        Hundreds
#> 96       IDR  Business    0.067941        6      false        Hundreds
#> 97       IDR  Business    0.067741        6      false        Hundreds
#> 98       IDR  Business    0.067644        6      false        Hundreds
#> 99       IDR  Business    0.067927        6      false        Hundreds
#> 100      IDR  Business    0.067907        6      false        Hundreds
#> 101      GBP  Business   13.612800        4      false           Units
#> 102      GBP  Business   13.664000        4      false           Units
#> 103      GBP  Business   13.695600        4      false           Units
#> 104      GBP  Business   13.707500        4      false           Units
#> 105      GBP  Business   13.740900        4      false           Units
#> 106      CNY  Business  150.720000        2      false        Hundreds
#> 107      CNY  Business  151.630000        2      false        Hundreds
#> 108      CNY  Business  151.880000        2      false        Hundreds
#> 109      CNY  Business  151.860000        2      false        Hundreds
#> 110      CNY  Business  152.520000        2      false        Hundreds
#> 111      BRL  Business    2.131100        4      false           Units
#> 112      BRL  Business    2.099500        4      false           Units
#> 113      BRL  Business    2.086900        4      false           Units
#> 114      BRL  Business    2.097000        4      false           Units
#> 115      BRL  Business    2.094700        4      false           Units
#> 116      VND  Business    0.043300        4      false        Hundreds
#> 117      VND  Business    0.043400        4      false        Hundreds
#> 118      VND  Business    0.043300        4      false        Hundreds
#> 119      VND  Business    0.043300        4      false        Hundreds
#> 120      VND  Business    0.043400        4      false        Hundreds
#> 121      BYN  Business    3.333800        4      false           Units
#> 122      BYN  Business    3.353700        4      false           Units
#> 123      BYN  Business    3.360100        4      false           Units
#> 124      BYN  Business    3.358600        4      false           Units
#> 125      BYN  Business    3.374600        4      false           Units
#> 126      TWI  Business  132.650000        2       true           Units
#> 127      TWI  Business  132.900000        2       true           Units
#> 128      TWI  Business  133.020000        2       true           Units
#> 129      TWI  Business  133.560000        2       true           Units
#> 130      TWI  Business  133.840000        2       true           Units
#> 131      SGD  Business    8.019000        4      false           Units
#> 132      SGD  Business    8.046400        4      false           Units
#> 133      SGD  Business    8.069000        4      false           Units
#> 134      SGD  Business    8.083700        4      false           Units
#> 135      SGD  Business    8.108200        4      false           Units
#> 136      PKR  Business    3.922000        3      false        Hundreds
#> 137      PKR  Business    3.943000        3      false        Hundreds
#> 138      PKR  Business    3.949000        3      false        Hundreds
#> 139      PKR  Business    3.946000        3      false        Hundreds
#> 140      PKR  Business    3.968000        3      false        Hundreds
#> 141      MXN  Business   65.850000        2      false        Hundreds
#> 142      MXN  Business   65.170000        2      false        Hundreds
#> 143      MXN  Business   64.700000        2      false        Hundreds
#> 144      MXN  Business   64.870000        2      false        Hundreds
#> 145      MXN  Business   63.660000        2      false        Hundreds
#> 146      ILS  Business    2.937500        4      false           Units
#> 147      ILS  Business    2.918600        4      false           Units
#> 148      ILS  Business    2.911400        4      false           Units
#> 149      ILS  Business    2.908000        4      false           Units
#> 150      ILS  Business    2.928000        4      false           Units
#> 151      I44  Business  120.660000        2       true           Units
#> 152      I44  Business  120.960000        2       true           Units
#> 153      I44  Business  121.080000        2       true           Units
#> 154      I44  Business  121.470000        2       true           Units
#> 155      I44  Business  121.770000        2       true           Units
#> 156      EUR  Business   11.626000        4      false           Units
#> 157      EUR  Business   11.674500        4      false           Units
#> 158      EUR  Business   11.696000        4      false           Units
#> 159      EUR  Business   11.737500        4      false           Units
#> 160      EUR  Business   11.765000        4      false           Units
#> 161      CHF  Business 1195.480000        2      false        Hundreds
#> 162      CHF  Business 1202.070000        2      false        Hundreds
#> 163      CHF  Business 1206.640000        2      false        Hundreds
#> 164      CHF  Business 1209.550000        2      false        Hundreds
#> 165      CHF  Business 1215.390000        2      false        Hundreds
#> 166      HRK  Business  138.580000        2      false        Hundreds
#> 167      HRK  Business  139.160000        2      false        Hundreds
#> 168      HRK  Business  138.650000        2      false        Hundreds
#> 169      HRK  Business  139.990000        2      false        Hundreds
#> 170      HRK  Business  139.510000        2      false        Hundreds
#> 171      BGN  Business  594.440000        2      false        Hundreds
#> 172      BGN  Business  596.920000        2      false        Hundreds
#> 173      BGN  Business  598.020000        2      false        Hundreds
#> 174      BGN  Business  600.140000        2      false        Hundreds
#> 175      BGN  Business  601.540000        2      false        Hundreds
#> 176      ISK  Business    7.710000        2      false        Hundreds
#> 177      ISK  Business    7.750000        2      false        Hundreds
#> 178      ISK  Business    7.770000        2      false        Hundreds
#> 179      ISK  Business    7.810000        2      false        Hundreds
#> 180      ISK  Business    7.830000        2      false        Hundreds
#> 181      ZAR  Business    0.575400        4      false           Units
#> 182      ZAR  Business    0.577600        4      false           Units
#> 183      ZAR  Business    0.576900        4      false           Units
#> 184      ZAR  Business    0.576900        4      false           Units
#> 185      ZAR  Business    0.574500        4      false           Units
#> 186      TWD  Business   33.675000        3      false        Hundreds
#> 187      TWD  Business   33.753000        3      false        Hundreds
#> 188      TWD  Business   33.861000        3      false        Hundreds
#> 189      TWD  Business   33.914000        3      false        Hundreds
#> 190      TWD  Business   33.960000        3      false        Hundreds
#> 191      SEK  Business  100.590000        2      false        Hundreds
#> 192      SEK  Business  100.340000        2      false        Hundreds
#> 193      SEK  Business  100.140000        2      false        Hundreds
#> 194      SEK  Business  100.860000        2      false        Hundreds
#> 195      SEK  Business  100.870000        2      false        Hundreds
#> 196      PHP  Business   19.198000        3      false        Hundreds
#> 197      PHP  Business   19.277000        3      false        Hundreds
#> 198      PHP  Business   19.212000        3      false        Hundreds
#> 199      PHP  Business   19.202000        3      false        Hundreds
#> 200      PHP  Business   19.170000        3      false        Hundreds
#> 201      KRW  Business    0.788900        4      false        Hundreds
#> 202      KRW  Business    0.788000        4      false        Hundreds
#> 203      KRW  Business    0.794600        4      false        Hundreds
#> 204      KRW  Business    0.799900        4      false        Hundreds
#> 205      KRW  Business    0.800800        4      false        Hundreds
#>      date_from    date_to       date
#> 1   2024-04-15 2024-04-15 2024-04-15
#> 2   2024-04-16 2024-04-16 2024-04-16
#> 3   2024-04-17 2024-04-17 2024-04-17
#> 4   2024-04-18 2024-04-18 2024-04-18
#> 5   2024-04-19 2024-04-19 2024-04-19
#> 6   2024-04-15 2024-04-15 2024-04-15
#> 7   2024-04-16 2024-04-16 2024-04-16
#> 8   2024-04-17 2024-04-17 2024-04-17
#> 9   2024-04-18 2024-04-18 2024-04-18
#> 10  2024-04-19 2024-04-19 2024-04-19
#> 11  2022-02-23 2022-02-23 2022-02-23
#> 12  2022-02-24 2022-02-24 2022-02-24
#> 13  2022-02-25 2022-02-25 2022-02-25
#> 14  2022-02-28 2022-02-28 2022-02-28
#> 15  2022-03-01 2022-03-01 2022-03-01
#> 16  2024-04-15 2024-04-15 2024-04-15
#> 17  2024-04-16 2024-04-16 2024-04-16
#> 18  2024-04-17 2024-04-17 2024-04-17
#> 19  2024-04-18 2024-04-18 2024-04-18
#> 20  2024-04-19 2024-04-19 2024-04-19
#> 21  2024-04-15 2024-04-15 2024-04-15
#> 22  2024-04-16 2024-04-16 2024-04-16
#> 23  2024-04-17 2024-04-17 2024-04-17
#> 24  2024-04-18 2024-04-18 2024-04-18
#> 25  2024-04-19 2024-04-19 2024-04-19
#> 26  2024-04-15 2024-04-15 2024-04-15
#> 27  2024-04-16 2024-04-16 2024-04-16
#> 28  2024-04-17 2024-04-17 2024-04-17
#> 29  2024-04-18 2024-04-18 2024-04-18
#> 30  2024-04-19 2024-04-19 2024-04-19
#> 31  2024-04-15 2024-04-15 2024-04-15
#> 32  2024-04-16 2024-04-16 2024-04-16
#> 33  2024-04-17 2024-04-17 2024-04-17
#> 34  2024-04-18 2024-04-18 2024-04-18
#> 35  2024-04-19 2024-04-19 2024-04-19
#> 36  2024-04-15 2024-04-15 2024-04-15
#> 37  2024-04-16 2024-04-16 2024-04-16
#> 38  2024-04-17 2024-04-17 2024-04-17
#> 39  2024-04-18 2024-04-18 2024-04-18
#> 40  2024-04-19 2024-04-19 2024-04-19
#> 41  2024-04-15 2024-04-15 2024-04-15
#> 42  2024-04-16 2024-04-16 2024-04-16
#> 43  2024-04-17 2024-04-17 2024-04-17
#> 44  2024-04-18 2024-04-18 2024-04-18
#> 45  2024-04-19 2024-04-19 2024-04-19
#> 46  2024-04-15 2024-04-15 2024-04-15
#> 47  2024-04-16 2024-04-16 2024-04-16
#> 48  2024-04-17 2024-04-17 2024-04-17
#> 49  2024-04-18 2024-04-18 2024-04-18
#> 50  2024-04-19 2024-04-19 2024-04-19
#> 51  2024-04-15 2024-04-15 2024-04-15
#> 52  2024-04-16 2024-04-16 2024-04-16
#> 53  2024-04-17 2024-04-17 2024-04-17
#> 54  2024-04-18 2024-04-18 2024-04-18
#> 55  2024-04-19 2024-04-19 2024-04-19
#> 56  2024-04-15 2024-04-15 2024-04-15
#> 57  2024-04-16 2024-04-16 2024-04-16
#> 58  2024-04-17 2024-04-17 2024-04-17
#> 59  2024-04-18 2024-04-18 2024-04-18
#> 60  2024-04-19 2024-04-19 2024-04-19
#> 61  2024-04-15 2024-04-15 2024-04-15
#> 62  2024-04-16 2024-04-16 2024-04-16
#> 63  2024-04-17 2024-04-17 2024-04-17
#> 64  2024-04-18 2024-04-18 2024-04-18
#> 65  2024-04-19 2024-04-19 2024-04-19
#> 66  2024-04-15 2024-04-15 2024-04-15
#> 67  2024-04-16 2024-04-16 2024-04-16
#> 68  2024-04-17 2024-04-17 2024-04-17
#> 69  2024-04-18 2024-04-18 2024-04-18
#> 70  2024-04-19 2024-04-19 2024-04-19
#> 71  2024-04-15 2024-04-15 2024-04-15
#> 72  2024-04-16 2024-04-16 2024-04-16
#> 73  2024-04-17 2024-04-17 2024-04-17
#> 74  2024-04-18 2024-04-18 2024-04-18
#> 75  2024-04-19 2024-04-19 2024-04-19
#> 76  2024-04-15 2024-04-15 2024-04-15
#> 77  2024-04-16 2024-04-16 2024-04-16
#> 78  2024-04-17 2024-04-17 2024-04-17
#> 79  2024-04-18 2024-04-18 2024-04-18
#> 80  2024-04-19 2024-04-19 2024-04-19
#> 81  2024-04-15 2024-04-15 2024-04-15
#> 82  2024-04-16 2024-04-16 2024-04-16
#> 83  2024-04-17 2024-04-17 2024-04-17
#> 84  2024-04-18 2024-04-18 2024-04-18
#> 85  2024-04-19 2024-04-19 2024-04-19
#> 86  2024-04-15 2024-04-15 2024-04-15
#> 87  2024-04-16 2024-04-16 2024-04-16
#> 88  2024-04-17 2024-04-17 2024-04-17
#> 89  2024-04-18 2024-04-18 2024-04-18
#> 90  2024-04-19 2024-04-19 2024-04-19
#> 91  2024-04-15 2024-04-15 2024-04-15
#> 92  2024-04-16 2024-04-16 2024-04-16
#> 93  2024-04-17 2024-04-17 2024-04-17
#> 94  2024-04-18 2024-04-18 2024-04-18
#> 95  2024-04-19 2024-04-19 2024-04-19
#> 96  2024-04-15 2024-04-15 2024-04-15
#> 97  2024-04-16 2024-04-16 2024-04-16
#> 98  2024-04-17 2024-04-17 2024-04-17
#> 99  2024-04-18 2024-04-18 2024-04-18
#> 100 2024-04-19 2024-04-19 2024-04-19
#> 101 2024-04-15 2024-04-15 2024-04-15
#> 102 2024-04-16 2024-04-16 2024-04-16
#> 103 2024-04-17 2024-04-17 2024-04-17
#> 104 2024-04-18 2024-04-18 2024-04-18
#> 105 2024-04-19 2024-04-19 2024-04-19
#> 106 2024-04-15 2024-04-15 2024-04-15
#> 107 2024-04-16 2024-04-16 2024-04-16
#> 108 2024-04-17 2024-04-17 2024-04-17
#> 109 2024-04-18 2024-04-18 2024-04-18
#> 110 2024-04-19 2024-04-19 2024-04-19
#> 111 2024-04-15 2024-04-15 2024-04-15
#> 112 2024-04-16 2024-04-16 2024-04-16
#> 113 2024-04-17 2024-04-17 2024-04-17
#> 114 2024-04-18 2024-04-18 2024-04-18
#> 115 2024-04-19 2024-04-19 2024-04-19
#> 116 2024-04-15 2024-04-15 2024-04-15
#> 117 2024-04-16 2024-04-16 2024-04-16
#> 118 2024-04-17 2024-04-17 2024-04-17
#> 119 2024-04-18 2024-04-18 2024-04-18
#> 120 2024-04-19 2024-04-19 2024-04-19
#> 121 2024-04-15 2024-04-15 2024-04-15
#> 122 2024-04-16 2024-04-16 2024-04-16
#> 123 2024-04-17 2024-04-17 2024-04-17
#> 124 2024-04-18 2024-04-18 2024-04-18
#> 125 2024-04-19 2024-04-19 2024-04-19
#> 126 2024-04-15 2024-04-15 2024-04-15
#> 127 2024-04-16 2024-04-16 2024-04-16
#> 128 2024-04-17 2024-04-17 2024-04-17
#> 129 2024-04-18 2024-04-18 2024-04-18
#> 130 2024-04-19 2024-04-19 2024-04-19
#> 131 2024-04-15 2024-04-15 2024-04-15
#> 132 2024-04-16 2024-04-16 2024-04-16
#> 133 2024-04-17 2024-04-17 2024-04-17
#> 134 2024-04-18 2024-04-18 2024-04-18
#> 135 2024-04-19 2024-04-19 2024-04-19
#> 136 2024-04-15 2024-04-15 2024-04-15
#> 137 2024-04-16 2024-04-16 2024-04-16
#> 138 2024-04-17 2024-04-17 2024-04-17
#> 139 2024-04-18 2024-04-18 2024-04-18
#> 140 2024-04-19 2024-04-19 2024-04-19
#> 141 2024-04-15 2024-04-15 2024-04-15
#> 142 2024-04-16 2024-04-16 2024-04-16
#> 143 2024-04-17 2024-04-17 2024-04-17
#> 144 2024-04-18 2024-04-18 2024-04-18
#> 145 2024-04-19 2024-04-19 2024-04-19
#> 146 2024-04-15 2024-04-15 2024-04-15
#> 147 2024-04-16 2024-04-16 2024-04-16
#> 148 2024-04-17 2024-04-17 2024-04-17
#> 149 2024-04-18 2024-04-18 2024-04-18
#> 150 2024-04-19 2024-04-19 2024-04-19
#> 151 2024-04-15 2024-04-15 2024-04-15
#> 152 2024-04-16 2024-04-16 2024-04-16
#> 153 2024-04-17 2024-04-17 2024-04-17
#> 154 2024-04-18 2024-04-18 2024-04-18
#> 155 2024-04-19 2024-04-19 2024-04-19
#> 156 2024-04-15 2024-04-15 2024-04-15
#> 157 2024-04-16 2024-04-16 2024-04-16
#> 158 2024-04-17 2024-04-17 2024-04-17
#> 159 2024-04-18 2024-04-18 2024-04-18
#> 160 2024-04-19 2024-04-19 2024-04-19
#> 161 2024-04-15 2024-04-15 2024-04-15
#> 162 2024-04-16 2024-04-16 2024-04-16
#> 163 2024-04-17 2024-04-17 2024-04-17
#> 164 2024-04-18 2024-04-18 2024-04-18
#> 165 2024-04-19 2024-04-19 2024-04-19
#> 166 2022-12-23 2022-12-23 2022-12-23
#> 167 2022-12-27 2022-12-27 2022-12-27
#> 168 2022-12-28 2022-12-28 2022-12-28
#> 169 2022-12-29 2022-12-29 2022-12-29
#> 170 2022-12-30 2022-12-30 2022-12-30
#> 171 2024-04-15 2024-04-15 2024-04-15
#> 172 2024-04-16 2024-04-16 2024-04-16
#> 173 2024-04-17 2024-04-17 2024-04-17
#> 174 2024-04-18 2024-04-18 2024-04-18
#> 175 2024-04-19 2024-04-19 2024-04-19
#> 176 2024-04-15 2024-04-15 2024-04-15
#> 177 2024-04-16 2024-04-16 2024-04-16
#> 178 2024-04-17 2024-04-17 2024-04-17
#> 179 2024-04-18 2024-04-18 2024-04-18
#> 180 2024-04-19 2024-04-19 2024-04-19
#> 181 2024-04-15 2024-04-15 2024-04-15
#> 182 2024-04-16 2024-04-16 2024-04-16
#> 183 2024-04-17 2024-04-17 2024-04-17
#> 184 2024-04-18 2024-04-18 2024-04-18
#> 185 2024-04-19 2024-04-19 2024-04-19
#> 186 2024-04-15 2024-04-15 2024-04-15
#> 187 2024-04-16 2024-04-16 2024-04-16
#> 188 2024-04-17 2024-04-17 2024-04-17
#> 189 2024-04-18 2024-04-18 2024-04-18
#> 190 2024-04-19 2024-04-19 2024-04-19
#> 191 2024-04-15 2024-04-15 2024-04-15
#> 192 2024-04-16 2024-04-16 2024-04-16
#> 193 2024-04-17 2024-04-17 2024-04-17
#> 194 2024-04-18 2024-04-18 2024-04-18
#> 195 2024-04-19 2024-04-19 2024-04-19
#> 196 2024-04-15 2024-04-15 2024-04-15
#> 197 2024-04-16 2024-04-16 2024-04-16
#> 198 2024-04-17 2024-04-17 2024-04-17
#> 199 2024-04-18 2024-04-18 2024-04-18
#> 200 2024-04-19 2024-04-19 2024-04-19
#> 201 2024-04-15 2024-04-15 2024-04-15
#> 202 2024-04-16 2024-04-16 2024-04-16
#> 203 2024-04-17 2024-04-17 2024-04-17
#> 204 2024-04-18 2024-04-18 2024-04-18
#> 205 2024-04-19 2024-04-19 2024-04-19