Get exchange rate

Fetch exchange rates from Norges Bank.

Usage

get_exchange_rate(
  currency,
  frequency = c("daily", "monthly", "annual"),
  n_obs = 10,
  start = NULL,
  end = NULL,
  simplify = TRUE,
  raw_response = FALSE
)

Arguments

currency

The currency to fetch data for.

frequency

Frequency of data. Can be daily, monthly or annual. Defaults to daily.

n_obs

The number of observations to fetch. See details.

start

The first date to fetch data for.

end

The last date to fetch data for.

simplify

If TRUE (default), a single data.frame is returned. Ignored if raw_response is set to TRUE.

raw_response

If TRUE a list of class noAPI is returned, including the raw httr2_response.

Value

data.frame or list

Details

Currency must be formatted as a three letter ISO 3166 code. A list of available country codes can be found at Norges Bank. You can fetch observations for several currencies by passing a character vector. If you want to include all currencies, you can set the currency argument to ALL. Note that the number of observations is limited to 20 when retrieving all currencies.

By default, the 10 latest observations are returned. You can adjust the number of observations by setting n_obs to a positive integer. You can also fetch exchange rates for a specific time period by setting the arguments start and end. Both arguments accept a date in the format %Y-%m-%d (YYYY-MM-DD).

The function returns a data.frame by default. If you prefer the the output as a list you can set simplify to FALSE. This can be useful to keep programmatically track of failed queries. If you set raw_response to TRUE, the raw response from the API will be returned together with the parsed response. Note that the response will then be returned silently.

Examples

# Get latest daily exchange rates
get_exchange_rate('EUR')
#>    currency frequency exch_rate decimals calculated unit_multiplier  date_from
#> 1       EUR  Business   11.1935        4      false           Units 2023-03-07
#> 2       EUR  Business   11.2205        4      false           Units 2023-03-08
#> 3       EUR  Business   11.2715        4      false           Units 2023-03-09
#> 4       EUR  Business   11.2900        4      false           Units 2023-03-10
#> 5       EUR  Business   11.3893        4      false           Units 2023-03-13
#> 6       EUR  Business   11.3350        4      false           Units 2023-03-14
#> 7       EUR  Business   11.3400        4      false           Units 2023-03-15
#> 8       EUR  Business   11.4487        4      false           Units 2023-03-16
#> 9       EUR  Business   11.4218        4      false           Units 2023-03-17
#> 10      EUR  Business   11.4210        4      false           Units 2023-03-20
#>       date_to       date
#> 1  2023-03-07 2023-03-07
#> 2  2023-03-08 2023-03-08
#> 3  2023-03-09 2023-03-09
#> 4  2023-03-10 2023-03-10
#> 5  2023-03-13 2023-03-13
#> 6  2023-03-14 2023-03-14
#> 7  2023-03-15 2023-03-15
#> 8  2023-03-16 2023-03-16
#> 9  2023-03-17 2023-03-17
#> 10 2023-03-20 2023-03-20
# Get monthly exchange rates for multiple currencies
get_exchange_rate(c('EUR', 'USD'), frequency = 'monthly')
#>    currency frequency exch_rate decimals calculated unit_multiplier  date_from
#> 1       EUR   Monthly   10.1373        4      false           Units 2022-05-01
#> 2       EUR   Monthly   10.3072        4      false           Units 2022-06-01
#> 3       EUR   Monthly   10.1823        4      false           Units 2022-07-01
#> 4       EUR   Monthly    9.8309        4      false           Units 2022-08-01
#> 5       EUR   Monthly   10.1697        4      false           Units 2022-09-01
#> 6       EUR   Monthly   10.3919        4      false           Units 2022-10-01
#> 7       EUR   Monthly   10.3357        4      false           Units 2022-11-01
#> 8       EUR   Monthly   10.4480        4      false           Units 2022-12-01
#> 9       EUR   Monthly   10.7149        4      false           Units 2023-01-01
#> 10      EUR   Monthly   10.9529        4      false           Units 2023-02-01
#> 11      USD   Monthly    9.5875        4      false           Units 2022-05-01
#> 12      USD   Monthly    9.7645        4      false           Units 2022-06-01
#> 13      USD   Monthly   10.0042        4      false           Units 2022-07-01
#> 14      USD   Monthly    9.7068        4      false           Units 2022-08-01
#> 15      USD   Monthly   10.2735        4      false           Units 2022-09-01
#> 16      USD   Monthly   10.5775        4      false           Units 2022-10-01
#> 17      USD   Monthly   10.1355        4      false           Units 2022-11-01
#> 18      USD   Monthly    9.8672        4      false           Units 2022-12-01
#> 19      USD   Monthly    9.9506        4      false           Units 2023-01-01
#> 20      USD   Monthly   10.2231        4      false           Units 2023-02-01
#>       date_to    date
#> 1  2022-05-31 2022-05
#> 2  2022-06-30 2022-06
#> 3  2022-07-31 2022-07
#> 4  2022-08-31 2022-08
#> 5  2022-09-30 2022-09
#> 6  2022-10-31 2022-10
#> 7  2022-11-30 2022-11
#> 8  2022-12-31 2022-12
#> 9  2023-01-31 2023-01
#> 10 2023-02-28 2023-02
#> 11 2022-05-31 2022-05
#> 12 2022-06-30 2022-06
#> 13 2022-07-31 2022-07
#> 14 2022-08-31 2022-08
#> 15 2022-09-30 2022-09
#> 16 2022-10-31 2022-10
#> 17 2022-11-30 2022-11
#> 18 2022-12-31 2022-12
#> 19 2023-01-31 2023-01
#> 20 2023-02-28 2023-02
# Get last 5 observations for all currencies
get_exchange_rate('ALL', n_obs = 5)
#>     currency frequency   exch_rate decimals calculated unit_multiplier
#> 1        RUB  Business   11.040000        3      false        Hundreds
#> 2        RUB  Business   10.539000        3      false        Hundreds
#> 3        RUB  Business   10.777000        3      false        Hundreds
#> 4        RUB  Business    8.613000        3      false        Hundreds
#> 5        RUB  Business    8.413000        3      false        Hundreds
#> 6        HRK  Business  138.580000        2      false        Hundreds
#> 7        HRK  Business  139.160000        2      false        Hundreds
#> 8        HRK  Business  138.650000        2      false        Hundreds
#> 9        HRK  Business  139.990000        2      false        Hundreds
#> 10       HRK  Business  139.510000        2      false        Hundreds
#> 11       AUD  Business    7.061400        4      false           Units
#> 12       AUD  Business    7.137500        4      false           Units
#> 13       AUD  Business    7.186000        4      false           Units
#> 14       AUD  Business    7.195300        4      false           Units
#> 15       AUD  Business    7.148000        4      false           Units
#> 16       BGN  Business  579.560000        2      false        Hundreds
#> 17       BGN  Business  579.810000        2      false        Hundreds
#> 18       BGN  Business  585.370000        2      false        Hundreds
#> 19       BGN  Business  584.000000        2      false        Hundreds
#> 20       BGN  Business  583.960000        2      false        Hundreds
#> 21       BRL  Business    2.019200        4      false           Units
#> 22       BRL  Business    2.036900        4      false           Units
#> 23       BRL  Business    2.044200        4      false           Units
#> 24       BRL  Business    2.044300        4      false           Units
#> 25       BRL  Business    2.032800        4      false           Units
#> 26       CAD  Business    7.716700        4      false           Units
#> 27       CAD  Business    7.818500        4      false           Units
#> 28       CAD  Business    7.860400        4      false           Units
#> 29       CAD  Business    7.831700        4      false           Units
#> 30       CAD  Business    7.785800        4      false           Units
#> 31       CHF  Business 1158.520000        2      false        Hundreds
#> 32       CHF  Business 1165.830000        2      false        Hundreds
#> 33       CHF  Business 1165.860000        2      false        Hundreds
#> 34       CHF  Business 1158.630000        2      false        Hundreds
#> 35       CHF  Business 1150.380000        2      false        Hundreds
#> 36       CNY  Business  153.680000        2      false        Hundreds
#> 37       CNY  Business  155.630000        2      false        Hundreds
#> 38       CNY  Business  156.730000        2      false        Hundreds
#> 39       CNY  Business  156.210000        2      false        Hundreds
#> 40       CNY  Business  154.880000        2      false        Hundreds
#> 41       CZK  Business   47.644000        3      false        Hundreds
#> 42       CZK  Business   47.364000        3      false        Hundreds
#> 43       CZK  Business   47.681000        3      false        Hundreds
#> 44       CZK  Business   47.694000        3      false        Hundreds
#> 45       CZK  Business   47.611000        3      false        Hundreds
#> 46       DKK  Business  152.260000        2      false        Hundreds
#> 47       DKK  Business  152.330000        2      false        Hundreds
#> 48       DKK  Business  153.720000        2      false        Hundreds
#> 49       DKK  Business  153.420000        2      false        Hundreds
#> 50       DKK  Business  153.400000        2      false        Hundreds
#> 51       EUR  Business   11.335000        4      false           Units
#> 52       EUR  Business   11.340000        4      false           Units
#> 53       EUR  Business   11.448700        4      false           Units
#> 54       EUR  Business   11.421800        4      false           Units
#> 55       EUR  Business   11.421000        4      false           Units
#> 56       GBP  Business   12.872800        4      false           Units
#> 57       GBP  Business   12.998200        4      false           Units
#> 58       GBP  Business   13.036600        4      false           Units
#> 59       GBP  Business   13.055300        4      false           Units
#> 60       GBP  Business   13.043600        4      false           Units
#> 61       HKD  Business    1.345400        4      false           Units
#> 62       HKD  Business    1.369600        4      false           Units
#> 63       HKD  Business    1.376600        4      false           Units
#> 64       HKD  Business    1.369700        4      false           Units
#> 65       HKD  Business    1.359000        4      false           Units
#> 66       HUF  Business    2.895300        4      false        Hundreds
#> 67       HUF  Business    2.852800        4      false        Hundreds
#> 68       HUF  Business    2.887900        4      false        Hundreds
#> 69       HUF  Business    2.884800        4      false        Hundreds
#> 70       HUF  Business    2.860900        4      false        Hundreds
#> 71       I44  Business  120.060000        2       true           Units
#> 72       I44  Business  120.910000        2       true           Units
#> 73       I44  Business  121.840000        2       true           Units
#> 74       I44  Business  121.480000        2       true           Units
#> 75       I44  Business  121.140000        2       true           Units
#> 76       IDR  Business    0.068726        6      false        Hundreds
#> 77       IDR  Business    0.069347        6      false        Hundreds
#> 78       IDR  Business    0.069958        6      false        Hundreds
#> 79       IDR  Business    0.069935        6      false        Hundreds
#> 80       IDR  Business    0.069414        6      false        Hundreds
#> 81       ILS  Business    2.911100        4      false           Units
#> 82       ILS  Business    2.958500        4      false           Units
#> 83       ILS  Business    2.943900        4      false           Units
#> 84       ILS  Business    2.924400        4      false           Units
#> 85       ILS  Business    2.901800        4      false           Units
#> 86       INR  Business   12.818000        3      false        Hundreds
#> 87       INR  Business   12.988000        3      false        Hundreds
#> 88       INR  Business   13.050000        3      false        Hundreds
#> 89       INR  Business   13.011000        3      false        Hundreds
#> 90       INR  Business   12.907000        3      false        Hundreds
#> 91       JPY  Business    7.866600        4      false        Hundreds
#> 92       JPY  Business    8.128400        4      false        Hundreds
#> 93       JPY  Business    8.167100        4      false        Hundreds
#> 94       JPY  Business    8.125300        4      false        Hundreds
#> 95       JPY  Business    8.113200        4      false        Hundreds
#> 96       KRW  Business    0.808300        4      false        Hundreds
#> 97       KRW  Business    0.813600        4      false        Hundreds
#> 98       KRW  Business    0.824600        4      false        Hundreds
#> 99       KRW  Business    0.821800        4      false        Hundreds
#> 100      KRW  Business    0.814700        4      false        Hundreds
#> 101      MXN  Business   56.420000        2      false        Hundreds
#> 102      MXN  Business   56.780000        2      false        Hundreds
#> 103      MXN  Business   56.660000        2      false        Hundreds
#> 104      MXN  Business   57.160000        2      false        Hundreds
#> 105      MXN  Business   56.260000        2      false        Hundreds
#> 106      MYR  Business    2.352800        4      false           Units
#> 107      MYR  Business    2.397700        4      false           Units
#> 108      MYR  Business    2.399100        4      false           Units
#> 109      MYR  Business    2.397100        4      false           Units
#> 110      MYR  Business    2.375600        4      false           Units
#> 111      NZD  Business    6.590900        4      false           Units
#> 112      NZD  Business    6.668200        4      false           Units
#> 113      NZD  Business    6.675600        4      false           Units
#> 114      NZD  Business    6.713200        4      false           Units
#> 115      NZD  Business    6.676600        4      false           Units
#> 116      PHP  Business   19.168000        3      false        Hundreds
#> 117      PHP  Business   19.535000        3      false        Hundreds
#> 118      PHP  Business   19.675000        3      false        Hundreds
#> 119      PHP  Business   19.653000        3      false        Hundreds
#> 120      PHP  Business   19.596000        3      false        Hundreds
#> 121      PKR  Business    3.742000        3      false        Hundreds
#> 122      PKR  Business    3.803000        3      false        Hundreds
#> 123      PKR  Business    3.828000        3      false        Hundreds
#> 124      PKR  Business    3.816000        3      false        Hundreds
#> 125      PKR  Business    3.781000        3      false        Hundreds
#> 126      PLN  Business    2.418000        4      false           Units
#> 127      PLN  Business    2.412800        4      false           Units
#> 128      PLN  Business    2.437800        4      false           Units
#> 129      PLN  Business    2.427800        4      false           Units
#> 130      PLN  Business    2.427700        4      false           Units
#> 131      RON  Business  230.300000        2      false        Hundreds
#> 132      RON  Business  230.400000        2      false        Hundreds
#> 133      RON  Business  232.630000        2      false        Hundreds
#> 134      RON  Business  232.110000        2      false        Hundreds
#> 135      RON  Business  231.980000        2      false        Hundreds
#> 136      SEK  Business  100.420000        2      false        Hundreds
#> 137      SEK  Business  101.230000        2      false        Hundreds
#> 138      SEK  Business  102.420000        2      false        Hundreds
#> 139      SEK  Business  101.980000        2      false        Hundreds
#> 140      SEK  Business  102.370000        2      false        Hundreds
#> 141      SGD  Business    7.856300        4      false           Units
#> 142      SGD  Business    7.960700        4      false           Units
#> 143      SGD  Business    8.019500        4      false           Units
#> 144      SGD  Business    8.008600        4      false           Units
#> 145      SGD  Business    7.956700        4      false           Units
#> 146      THB  Business   30.627000        3      false        Hundreds
#> 147      THB  Business   31.024000        3      false        Hundreds
#> 148      THB  Business   31.421000        3      false        Hundreds
#> 149      THB  Business   31.512000        3      false        Hundreds
#> 150      THB  Business   31.279000        3      false        Hundreds
#> 151      TRY  Business   55.640000        2      false        Hundreds
#> 152      TRY  Business   56.620000        2      false        Hundreds
#> 153      TRY  Business   56.900000        2      false        Hundreds
#> 154      TRY  Business   56.580000        2      false        Hundreds
#> 155      TRY  Business   56.030000        2      false        Hundreds
#> 156      TWD  Business   34.611000        3      false        Hundreds
#> 157      TWD  Business   34.982000        3      false        Hundreds
#> 158      TWD  Business   35.319000        3      false        Hundreds
#> 159      TWD  Business   35.142000        3      false        Hundreds
#> 160      TWD  Business   34.897000        3      false        Hundreds
#> 161      TWI  Business  131.090000        2       true           Units
#> 162      TWI  Business  131.840000        2       true           Units
#> 163      TWI  Business  133.020000        2       true           Units
#> 164      TWI  Business  132.630000        2       true           Units
#> 165      TWI  Business  132.470000        2       true           Units
#> 166      USD  Business   10.557000        4      false           Units
#> 167      USD  Business   10.749800        4      false           Units
#> 168      USD  Business   10.805800        4      false           Units
#> 169      USD  Business   10.752000        4      false           Units
#> 170      USD  Business   10.656900        4      false           Units
#> 171      XDR  Business   14.128820        5       true           Units
#> 172      XDR  Business   14.309030        5       true           Units
#> 173      XDR  Business   14.402440        5       true           Units
#> 174      XDR  Business   14.351430        5       true           Units
#> 175      XDR  Business   14.279000        5       true           Units
#> 176      ZAR  Business    0.584500        4      false           Units
#> 177      ZAR  Business    0.584000        4      false           Units
#> 178      ZAR  Business    0.587800        4      false           Units
#> 179      ZAR  Business    0.585800        4      false           Units
#> 180      ZAR  Business    0.577200        4      false           Units
#> 181      BYN  Business    4.182500        4      false           Units
#> 182      BYN  Business    4.258900        4      false           Units
#> 183      BYN  Business    4.281000        4      false           Units
#> 184      BYN  Business    4.259700        4      false           Units
#> 185      BYN  Business    4.222100        4      false           Units
#> 186      BDT  Business   10.000000        2      false        Hundreds
#> 187      BDT  Business   10.190000        2      false        Hundreds
#> 188      BDT  Business   10.080000        2      false        Hundreds
#> 189      BDT  Business   10.030000        2      false        Hundreds
#> 190      BDT  Business   10.120000        2      false        Hundreds
#> 191      MMK  Business    0.502700        4      false        Hundreds
#> 192      MMK  Business    0.511900        4      false        Hundreds
#> 193      MMK  Business    0.514600        4      false        Hundreds
#> 194      MMK  Business    0.512000        4      false        Hundreds
#> 195      MMK  Business    0.507500        4      false        Hundreds
#> 196      ISK  Business    7.500000        2      false        Hundreds
#> 197      ISK  Business    7.540000        2      false        Hundreds
#> 198      ISK  Business    7.630000        2      false        Hundreds
#> 199      ISK  Business    7.650000        2      false        Hundreds
#> 200      ISK  Business    7.610000        2      false        Hundreds
#> 201      VND  Business    0.044800        4      false        Hundreds
#> 202      VND  Business    0.045600        4      false        Hundreds
#> 203      VND  Business    0.045800        4      false        Hundreds
#> 204      VND  Business    0.045600        4      false        Hundreds
#> 205      VND  Business    0.045200        4      false        Hundreds
#>      date_from    date_to       date
#> 1   2022-02-23 2022-02-23 2022-02-23
#> 2   2022-02-24 2022-02-24 2022-02-24
#> 3   2022-02-25 2022-02-25 2022-02-25
#> 4   2022-02-28 2022-02-28 2022-02-28
#> 5   2022-03-01 2022-03-01 2022-03-01
#> 6   2022-12-23 2022-12-23 2022-12-23
#> 7   2022-12-27 2022-12-27 2022-12-27
#> 8   2022-12-28 2022-12-28 2022-12-28
#> 9   2022-12-29 2022-12-29 2022-12-29
#> 10  2022-12-30 2022-12-30 2022-12-30
#> 11  2023-03-14 2023-03-14 2023-03-14
#> 12  2023-03-15 2023-03-15 2023-03-15
#> 13  2023-03-16 2023-03-16 2023-03-16
#> 14  2023-03-17 2023-03-17 2023-03-17
#> 15  2023-03-20 2023-03-20 2023-03-20
#> 16  2023-03-14 2023-03-14 2023-03-14
#> 17  2023-03-15 2023-03-15 2023-03-15
#> 18  2023-03-16 2023-03-16 2023-03-16
#> 19  2023-03-17 2023-03-17 2023-03-17
#> 20  2023-03-20 2023-03-20 2023-03-20
#> 21  2023-03-14 2023-03-14 2023-03-14
#> 22  2023-03-15 2023-03-15 2023-03-15
#> 23  2023-03-16 2023-03-16 2023-03-16
#> 24  2023-03-17 2023-03-17 2023-03-17
#> 25  2023-03-20 2023-03-20 2023-03-20
#> 26  2023-03-14 2023-03-14 2023-03-14
#> 27  2023-03-15 2023-03-15 2023-03-15
#> 28  2023-03-16 2023-03-16 2023-03-16
#> 29  2023-03-17 2023-03-17 2023-03-17
#> 30  2023-03-20 2023-03-20 2023-03-20
#> 31  2023-03-14 2023-03-14 2023-03-14
#> 32  2023-03-15 2023-03-15 2023-03-15
#> 33  2023-03-16 2023-03-16 2023-03-16
#> 34  2023-03-17 2023-03-17 2023-03-17
#> 35  2023-03-20 2023-03-20 2023-03-20
#> 36  2023-03-14 2023-03-14 2023-03-14
#> 37  2023-03-15 2023-03-15 2023-03-15
#> 38  2023-03-16 2023-03-16 2023-03-16
#> 39  2023-03-17 2023-03-17 2023-03-17
#> 40  2023-03-20 2023-03-20 2023-03-20
#> 41  2023-03-14 2023-03-14 2023-03-14
#> 42  2023-03-15 2023-03-15 2023-03-15
#> 43  2023-03-16 2023-03-16 2023-03-16
#> 44  2023-03-17 2023-03-17 2023-03-17
#> 45  2023-03-20 2023-03-20 2023-03-20
#> 46  2023-03-14 2023-03-14 2023-03-14
#> 47  2023-03-15 2023-03-15 2023-03-15
#> 48  2023-03-16 2023-03-16 2023-03-16
#> 49  2023-03-17 2023-03-17 2023-03-17
#> 50  2023-03-20 2023-03-20 2023-03-20
#> 51  2023-03-14 2023-03-14 2023-03-14
#> 52  2023-03-15 2023-03-15 2023-03-15
#> 53  2023-03-16 2023-03-16 2023-03-16
#> 54  2023-03-17 2023-03-17 2023-03-17
#> 55  2023-03-20 2023-03-20 2023-03-20
#> 56  2023-03-14 2023-03-14 2023-03-14
#> 57  2023-03-15 2023-03-15 2023-03-15
#> 58  2023-03-16 2023-03-16 2023-03-16
#> 59  2023-03-17 2023-03-17 2023-03-17
#> 60  2023-03-20 2023-03-20 2023-03-20
#> 61  2023-03-14 2023-03-14 2023-03-14
#> 62  2023-03-15 2023-03-15 2023-03-15
#> 63  2023-03-16 2023-03-16 2023-03-16
#> 64  2023-03-17 2023-03-17 2023-03-17
#> 65  2023-03-20 2023-03-20 2023-03-20
#> 66  2023-03-14 2023-03-14 2023-03-14
#> 67  2023-03-15 2023-03-15 2023-03-15
#> 68  2023-03-16 2023-03-16 2023-03-16
#> 69  2023-03-17 2023-03-17 2023-03-17
#> 70  2023-03-20 2023-03-20 2023-03-20
#> 71  2023-03-14 2023-03-14 2023-03-14
#> 72  2023-03-15 2023-03-15 2023-03-15
#> 73  2023-03-16 2023-03-16 2023-03-16
#> 74  2023-03-17 2023-03-17 2023-03-17
#> 75  2023-03-20 2023-03-20 2023-03-20
#> 76  2023-03-14 2023-03-14 2023-03-14
#> 77  2023-03-15 2023-03-15 2023-03-15
#> 78  2023-03-16 2023-03-16 2023-03-16
#> 79  2023-03-17 2023-03-17 2023-03-17
#> 80  2023-03-20 2023-03-20 2023-03-20
#> 81  2023-03-14 2023-03-14 2023-03-14
#> 82  2023-03-15 2023-03-15 2023-03-15
#> 83  2023-03-16 2023-03-16 2023-03-16
#> 84  2023-03-17 2023-03-17 2023-03-17
#> 85  2023-03-20 2023-03-20 2023-03-20
#> 86  2023-03-14 2023-03-14 2023-03-14
#> 87  2023-03-15 2023-03-15 2023-03-15
#> 88  2023-03-16 2023-03-16 2023-03-16
#> 89  2023-03-17 2023-03-17 2023-03-17
#> 90  2023-03-20 2023-03-20 2023-03-20
#> 91  2023-03-14 2023-03-14 2023-03-14
#> 92  2023-03-15 2023-03-15 2023-03-15
#> 93  2023-03-16 2023-03-16 2023-03-16
#> 94  2023-03-17 2023-03-17 2023-03-17
#> 95  2023-03-20 2023-03-20 2023-03-20
#> 96  2023-03-14 2023-03-14 2023-03-14
#> 97  2023-03-15 2023-03-15 2023-03-15
#> 98  2023-03-16 2023-03-16 2023-03-16
#> 99  2023-03-17 2023-03-17 2023-03-17
#> 100 2023-03-20 2023-03-20 2023-03-20
#> 101 2023-03-14 2023-03-14 2023-03-14
#> 102 2023-03-15 2023-03-15 2023-03-15
#> 103 2023-03-16 2023-03-16 2023-03-16
#> 104 2023-03-17 2023-03-17 2023-03-17
#> 105 2023-03-20 2023-03-20 2023-03-20
#> 106 2023-03-14 2023-03-14 2023-03-14
#> 107 2023-03-15 2023-03-15 2023-03-15
#> 108 2023-03-16 2023-03-16 2023-03-16
#> 109 2023-03-17 2023-03-17 2023-03-17
#> 110 2023-03-20 2023-03-20 2023-03-20
#> 111 2023-03-14 2023-03-14 2023-03-14
#> 112 2023-03-15 2023-03-15 2023-03-15
#> 113 2023-03-16 2023-03-16 2023-03-16
#> 114 2023-03-17 2023-03-17 2023-03-17
#> 115 2023-03-20 2023-03-20 2023-03-20
#> 116 2023-03-14 2023-03-14 2023-03-14
#> 117 2023-03-15 2023-03-15 2023-03-15
#> 118 2023-03-16 2023-03-16 2023-03-16
#> 119 2023-03-17 2023-03-17 2023-03-17
#> 120 2023-03-20 2023-03-20 2023-03-20
#> 121 2023-03-14 2023-03-14 2023-03-14
#> 122 2023-03-15 2023-03-15 2023-03-15
#> 123 2023-03-16 2023-03-16 2023-03-16
#> 124 2023-03-17 2023-03-17 2023-03-17
#> 125 2023-03-20 2023-03-20 2023-03-20
#> 126 2023-03-14 2023-03-14 2023-03-14
#> 127 2023-03-15 2023-03-15 2023-03-15
#> 128 2023-03-16 2023-03-16 2023-03-16
#> 129 2023-03-17 2023-03-17 2023-03-17
#> 130 2023-03-20 2023-03-20 2023-03-20
#> 131 2023-03-14 2023-03-14 2023-03-14
#> 132 2023-03-15 2023-03-15 2023-03-15
#> 133 2023-03-16 2023-03-16 2023-03-16
#> 134 2023-03-17 2023-03-17 2023-03-17
#> 135 2023-03-20 2023-03-20 2023-03-20
#> 136 2023-03-14 2023-03-14 2023-03-14
#> 137 2023-03-15 2023-03-15 2023-03-15
#> 138 2023-03-16 2023-03-16 2023-03-16
#> 139 2023-03-17 2023-03-17 2023-03-17
#> 140 2023-03-20 2023-03-20 2023-03-20
#> 141 2023-03-14 2023-03-14 2023-03-14
#> 142 2023-03-15 2023-03-15 2023-03-15
#> 143 2023-03-16 2023-03-16 2023-03-16
#> 144 2023-03-17 2023-03-17 2023-03-17
#> 145 2023-03-20 2023-03-20 2023-03-20
#> 146 2023-03-14 2023-03-14 2023-03-14
#> 147 2023-03-15 2023-03-15 2023-03-15
#> 148 2023-03-16 2023-03-16 2023-03-16
#> 149 2023-03-17 2023-03-17 2023-03-17
#> 150 2023-03-20 2023-03-20 2023-03-20
#> 151 2023-03-14 2023-03-14 2023-03-14
#> 152 2023-03-15 2023-03-15 2023-03-15
#> 153 2023-03-16 2023-03-16 2023-03-16
#> 154 2023-03-17 2023-03-17 2023-03-17
#> 155 2023-03-20 2023-03-20 2023-03-20
#> 156 2023-03-14 2023-03-14 2023-03-14
#> 157 2023-03-15 2023-03-15 2023-03-15
#> 158 2023-03-16 2023-03-16 2023-03-16
#> 159 2023-03-17 2023-03-17 2023-03-17
#> 160 2023-03-20 2023-03-20 2023-03-20
#> 161 2023-03-14 2023-03-14 2023-03-14
#> 162 2023-03-15 2023-03-15 2023-03-15
#> 163 2023-03-16 2023-03-16 2023-03-16
#> 164 2023-03-17 2023-03-17 2023-03-17
#> 165 2023-03-20 2023-03-20 2023-03-20
#> 166 2023-03-14 2023-03-14 2023-03-14
#> 167 2023-03-15 2023-03-15 2023-03-15
#> 168 2023-03-16 2023-03-16 2023-03-16
#> 169 2023-03-17 2023-03-17 2023-03-17
#> 170 2023-03-20 2023-03-20 2023-03-20
#> 171 2023-03-14 2023-03-14 2023-03-14
#> 172 2023-03-15 2023-03-15 2023-03-15
#> 173 2023-03-16 2023-03-16 2023-03-16
#> 174 2023-03-17 2023-03-17 2023-03-17
#> 175 2023-03-20 2023-03-20 2023-03-20
#> 176 2023-03-14 2023-03-14 2023-03-14
#> 177 2023-03-15 2023-03-15 2023-03-15
#> 178 2023-03-16 2023-03-16 2023-03-16
#> 179 2023-03-17 2023-03-17 2023-03-17
#> 180 2023-03-20 2023-03-20 2023-03-20
#> 181 2023-03-14 2023-03-14 2023-03-14
#> 182 2023-03-15 2023-03-15 2023-03-15
#> 183 2023-03-16 2023-03-16 2023-03-16
#> 184 2023-03-17 2023-03-17 2023-03-17
#> 185 2023-03-20 2023-03-20 2023-03-20
#> 186 2023-03-14 2023-03-14 2023-03-14
#> 187 2023-03-15 2023-03-15 2023-03-15
#> 188 2023-03-16 2023-03-16 2023-03-16
#> 189 2023-03-17 2023-03-17 2023-03-17
#> 190 2023-03-20 2023-03-20 2023-03-20
#> 191 2023-03-14 2023-03-14 2023-03-14
#> 192 2023-03-15 2023-03-15 2023-03-15
#> 193 2023-03-16 2023-03-16 2023-03-16
#> 194 2023-03-17 2023-03-17 2023-03-17
#> 195 2023-03-20 2023-03-20 2023-03-20
#> 196 2023-03-14 2023-03-14 2023-03-14
#> 197 2023-03-15 2023-03-15 2023-03-15
#> 198 2023-03-16 2023-03-16 2023-03-16
#> 199 2023-03-17 2023-03-17 2023-03-17
#> 200 2023-03-20 2023-03-20 2023-03-20
#> 201 2023-03-14 2023-03-14 2023-03-14
#> 202 2023-03-15 2023-03-15 2023-03-15
#> 203 2023-03-16 2023-03-16 2023-03-16
#> 204 2023-03-17 2023-03-17 2023-03-17
#> 205 2023-03-20 2023-03-20 2023-03-20