Get exchange rate

Fetch exchange rates from Norges Bank.

Usage

get_exchange_rate(
  currency,
  frequency = c("daily", "monthly", "annual"),
  n_obs = 10,
  start = NULL,
  end = NULL,
  raw_response = FALSE
)

Arguments

currency

The currency to fetch data for.

frequency

Frequency of data. Can be daily, monthly or annual. Defaults to daily.

n_obs

The number of observations to fetch. See details.

start

The first date to fetch data for.

end

The last date to fetch data for.

raw_response

If TRUE a list of class noAPI is returned, including the raw httr2_response.

Value

data.frame or list

Details

Currency must be formatted as a three letter ISO 3166 code. A list of available country codes can be found at Norges Bank. You can fetch observations for several currencies by passing a character vector. If you want to include all currencies, you can set the currency argument to ALL. Note that the number of observations is limited to 20 when retrieving all currencies.

By default, the 10 latest observations are returned. You can adjust the number of observations by setting n_obs to a positive integer. You can also fetch exchange rates for a specific time period by setting the arguments start and end. Both arguments accept a date in the format %Y-%m-%d (YYYY-MM-DD).

The function returns a data.frame by default. If you set raw_response to TRUE, the raw response from the API will be returned instead. Note that the response will be returned silently when raw_response is set to TRUE.

Examples

# Get latest daily exchange rates
get_exchange_rate('EUR')
#>    currency frequency exch_rate decimals calculated unit_multiplier  date_from
#> 1       EUR  Business   10.9123        4      false           Units 2023-02-09
#> 2       EUR  Business    10.871        4      false           Units 2023-02-10
#> 3       EUR  Business   10.8213        4      false           Units 2023-02-13
#> 4       EUR  Business   10.8778        4      false           Units 2023-02-14
#> 5       EUR  Business   10.9193        4      false           Units 2023-02-15
#> 6       EUR  Business   10.9513        4      false           Units 2023-02-16
#> 7       EUR  Business    11.015        4      false           Units 2023-02-17
#> 8       EUR  Business   10.9683        4      false           Units 2023-02-20
#> 9       EUR  Business   10.9468        4      false           Units 2023-02-21
#> 10      EUR  Business   10.9665        4      false           Units 2023-02-22
#>       date_to       date
#> 1  2023-02-09 2023-02-09
#> 2  2023-02-10 2023-02-10
#> 3  2023-02-13 2023-02-13
#> 4  2023-02-14 2023-02-14
#> 5  2023-02-15 2023-02-15
#> 6  2023-02-16 2023-02-16
#> 7  2023-02-17 2023-02-17
#> 8  2023-02-20 2023-02-20
#> 9  2023-02-21 2023-02-21
#> 10 2023-02-22 2023-02-22
# Get monthly exchange rates for multiple currencies
get_exchange_rate(c('EUR', 'USD'), frequency = 'monthly')
#>    currency frequency exch_rate decimals calculated unit_multiplier  date_from
#> 1       EUR   Monthly    9.6239        4      false           Units 2022-04-01
#> 2       EUR   Monthly   10.1373        4      false           Units 2022-05-01
#> 3       EUR   Monthly   10.3072        4      false           Units 2022-06-01
#> 4       EUR   Monthly   10.1823        4      false           Units 2022-07-01
#> 5       EUR   Monthly    9.8309        4      false           Units 2022-08-01
#> 6       EUR   Monthly   10.1697        4      false           Units 2022-09-01
#> 7       EUR   Monthly   10.3919        4      false           Units 2022-10-01
#> 8       EUR   Monthly   10.3357        4      false           Units 2022-11-01
#> 9       EUR   Monthly    10.448        4      false           Units 2022-12-01
#> 10      EUR   Monthly   10.7149        4      false           Units 2023-01-01
#> 11      USD   Monthly    8.9013        4      false           Units 2022-04-01
#> 12      USD   Monthly    9.5875        4      false           Units 2022-05-01
#> 13      USD   Monthly    9.7645        4      false           Units 2022-06-01
#> 14      USD   Monthly   10.0042        4      false           Units 2022-07-01
#> 15      USD   Monthly    9.7068        4      false           Units 2022-08-01
#> 16      USD   Monthly   10.2735        4      false           Units 2022-09-01
#> 17      USD   Monthly   10.5775        4      false           Units 2022-10-01
#> 18      USD   Monthly   10.1355        4      false           Units 2022-11-01
#> 19      USD   Monthly    9.8672        4      false           Units 2022-12-01
#> 20      USD   Monthly    9.9506        4      false           Units 2023-01-01
#>       date_to    date
#> 1  2022-04-30 2022-04
#> 2  2022-05-31 2022-05
#> 3  2022-06-30 2022-06
#> 4  2022-07-31 2022-07
#> 5  2022-08-31 2022-08
#> 6  2022-09-30 2022-09
#> 7  2022-10-31 2022-10
#> 8  2022-11-30 2022-11
#> 9  2022-12-31 2022-12
#> 10 2023-01-31 2023-01
#> 11 2022-04-30 2022-04
#> 12 2022-05-31 2022-05
#> 13 2022-06-30 2022-06
#> 14 2022-07-31 2022-07
#> 15 2022-08-31 2022-08
#> 16 2022-09-30 2022-09
#> 17 2022-10-31 2022-10
#> 18 2022-11-30 2022-11
#> 19 2022-12-31 2022-12
#> 20 2023-01-31 2023-01
# Get last 5 observations for all currencies
get_exchange_rate('ALL', n_obs = 5)
#>     currency frequency exch_rate decimals calculated unit_multiplier  date_from
#> 1        RUB  Business     11.04        3      false        Hundreds 2022-02-23
#> 2        RUB  Business    10.539        3      false        Hundreds 2022-02-24
#> 3        RUB  Business    10.777        3      false        Hundreds 2022-02-25
#> 4        RUB  Business     8.613        3      false        Hundreds 2022-02-28
#> 5        RUB  Business     8.413        3      false        Hundreds 2022-03-01
#> 6        HRK  Business    138.58        2      false        Hundreds 2022-12-23
#> 7        HRK  Business    139.16        2      false        Hundreds 2022-12-27
#> 8        HRK  Business    138.65        2      false        Hundreds 2022-12-28
#> 9        HRK  Business    139.99        2      false        Hundreds 2022-12-29
#> 10       HRK  Business    139.51        2      false        Hundreds 2022-12-30
#> 11       AUD  Business    7.0654        4      false           Units 2023-02-16
#> 12       AUD  Business     7.079        4      false           Units 2023-02-17
#> 13       AUD  Business    7.0965        4      false           Units 2023-02-20
#> 14       AUD  Business    7.0547        4      false           Units 2023-02-21
#> 15       AUD  Business    7.0406        4      false           Units 2023-02-22
#> 16       BGN  Business    559.94        2      false        Hundreds 2023-02-16
#> 17       BGN  Business     563.2        2      false        Hundreds 2023-02-17
#> 18       BGN  Business    560.81        2      false        Hundreds 2023-02-20
#> 19       BGN  Business    559.71        2      false        Hundreds 2023-02-21
#> 20       BGN  Business    560.72        2      false        Hundreds 2023-02-22
#> 21       BRL  Business    1.9525        4      false           Units 2023-02-16
#> 22       BRL  Business    1.9851        4      false           Units 2023-02-17
#> 23       BRL  Business    1.9884        4      false           Units 2023-02-20
#> 24       BRL  Business    1.9864        4      false           Units 2023-02-21
#> 25       BRL  Business    1.9936        4      false           Units 2023-02-22
#> 26       CAD  Business    7.6401        4      false           Units 2023-02-16
#> 27       CAD  Business    7.6674        4      false           Units 2023-02-17
#> 28       CAD  Business    7.6344        4      false           Units 2023-02-20
#> 29       CAD  Business    7.6279        4      false           Units 2023-02-21
#> 30       CAD  Business    7.6114        4      false           Units 2023-02-22
#> 31       CHF  Business   1109.22        2      false        Hundreds 2023-02-16
#> 32       CHF  Business   1113.08        2      false        Hundreds 2023-02-17
#> 33       CHF  Business   1112.29        2      false        Hundreds 2023-02-20
#> 34       CHF  Business   1111.01        2      false        Hundreds 2023-02-21
#> 35       CHF  Business   1111.09        2      false        Hundreds 2023-02-22
#> 36       CNY  Business     149.3        2      false        Hundreds 2023-02-16
#> 37       CNY  Business    150.72        2      false        Hundreds 2023-02-17
#> 38       CNY  Business    149.86        2      false        Hundreds 2023-02-20
#> 39       CNY  Business     149.2        2      false        Hundreds 2023-02-21
#> 40       CNY  Business    149.47        2      false        Hundreds 2023-02-22
#> 41       CZK  Business    46.237        3      false        Hundreds 2023-02-16
#> 42       CZK  Business    46.447        3      false        Hundreds 2023-02-17
#> 43       CZK  Business    46.293        3      false        Hundreds 2023-02-20
#> 44       CZK  Business    46.131        3      false        Hundreds 2023-02-21
#> 45       CZK  Business    46.333        3      false        Hundreds 2023-02-22
#> 46       DKK  Business    147.02        2      false        Hundreds 2023-02-16
#> 47       DKK  Business    147.92        2      false        Hundreds 2023-02-17
#> 48       DKK  Business     147.3        2      false        Hundreds 2023-02-20
#> 49       DKK  Business    147.02        2      false        Hundreds 2023-02-21
#> 50       DKK  Business    147.28        2      false        Hundreds 2023-02-22
#> 51       EUR  Business   10.9513        4      false           Units 2023-02-16
#> 52       EUR  Business    11.015        4      false           Units 2023-02-17
#> 53       EUR  Business   10.9683        4      false           Units 2023-02-20
#> 54       EUR  Business   10.9468        4      false           Units 2023-02-21
#> 55       EUR  Business   10.9665        4      false           Units 2023-02-22
#> 56       GBP  Business    12.321        4      false           Units 2023-02-16
#> 57       GBP  Business    12.392        4      false           Units 2023-02-17
#> 58       GBP  Business   12.3603        4      false           Units 2023-02-20
#> 59       GBP  Business   12.4502        4      false           Units 2023-02-21
#> 60       GBP  Business   12.4697        4      false           Units 2023-02-22
#> 61       HKD  Business     1.304        4      false           Units 2023-02-16
#> 62       HKD  Business    1.3216        4      false           Units 2023-02-17
#> 63       HKD  Business    1.3116        4      false           Units 2023-02-20
#> 64       HKD  Business    1.3087        4      false           Units 2023-02-21
#> 65       HKD  Business    1.3132        4      false           Units 2023-02-22
#> 66       HUF  Business     2.862        4      false        Hundreds 2023-02-16
#> 67       HUF  Business    2.8491        4      false        Hundreds 2023-02-17
#> 68       HUF  Business    2.8651        4      false        Hundreds 2023-02-20
#> 69       HUF  Business    2.8642        4      false        Hundreds 2023-02-21
#> 70       HUF  Business    2.8683        4      false        Hundreds 2023-02-22
#> 71       I44  Business    116.41        2       true           Units 2023-02-16
#> 72       I44  Business    117.22        2       true           Units 2023-02-17
#> 73       I44  Business    116.84        2       true           Units 2023-02-20
#> 74       I44  Business    116.65        2       true           Units 2023-02-21
#> 75       I44  Business    116.88        2       true           Units 2023-02-22
#> 76       IDR  Business  0.067628        6      false        Hundreds 2023-02-16
#> 77       IDR  Business  0.068159        6      false        Hundreds 2023-02-17
#> 78       IDR  Business  0.067795        6      false        Hundreds 2023-02-20
#> 79       IDR  Business  0.067594        6      false        Hundreds 2023-02-21
#> 80       IDR  Business  0.067786        6      false        Hundreds 2023-02-22
#> 81       ILS  Business     2.889        4      false           Units 2023-02-16
#> 82       ILS  Business    2.9102        4      false           Units 2023-02-17
#> 83       ILS  Business     2.884        4      false           Units 2023-02-20
#> 84       ILS  Business    2.8171        4      false           Units 2023-02-21
#> 85       ILS  Business    2.8221        4      false           Units 2023-02-22
#> 86       INR  Business     12.38        3      false        Hundreds 2023-02-16
#> 87       INR  Business    12.514        3      false        Hundreds 2023-02-17
#> 88       INR  Business    12.421        3      false        Hundreds 2023-02-20
#> 89       INR  Business      12.4        3      false        Hundreds 2023-02-21
#> 90       INR  Business    12.428        3      false        Hundreds 2023-02-22
#> 91       JPY  Business    7.6417        4      false        Hundreds 2023-02-16
#> 92       JPY  Business    7.6931        4      false        Hundreds 2023-02-17
#> 93       JPY  Business    7.6653        4      false        Hundreds 2023-02-20
#> 94       JPY  Business    7.6146        4      false        Hundreds 2023-02-21
#> 95       JPY  Business     7.656        4      false        Hundreds 2023-02-22
#> 96       KRW  Business    0.7956        4      false        Hundreds 2023-02-16
#> 97       KRW  Business    0.7963        4      false        Hundreds 2023-02-17
#> 98       KRW  Business    0.7923        4      false        Hundreds 2023-02-20
#> 99       KRW  Business    0.7892        4      false        Hundreds 2023-02-21
#> 100      KRW  Business    0.7907        4      false        Hundreds 2023-02-22
#> 101      MXN  Business     55.11        2      false        Hundreds 2023-02-16
#> 102      MXN  Business     55.83        2      false        Hundreds 2023-02-17
#> 103      MXN  Business      55.8        2      false        Hundreds 2023-02-20
#> 104      MXN  Business     55.83        2      false        Hundreds 2023-02-21
#> 105      MXN  Business     55.96        2      false        Hundreds 2023-02-22
#> 106      MYR  Business     2.324        4      false           Units 2023-02-16
#> 107      MYR  Business    2.3391        4      false           Units 2023-02-17
#> 108      MYR  Business    2.3193        4      false           Units 2023-02-20
#> 109      MYR  Business    2.3159        4      false           Units 2023-02-21
#> 110      MYR  Business    2.3197        4      false           Units 2023-02-22
#> 111      NZD  Business    6.4295        4      false           Units 2023-02-16
#> 112      NZD  Business    6.4329        4      false           Units 2023-02-17
#> 113      NZD  Business    6.4191        4      false           Units 2023-02-20
#> 114      NZD  Business    6.4046        4      false           Units 2023-02-21
#> 115      NZD  Business    6.4259        4      false           Units 2023-02-22
#> 116      PHP  Business    18.558        3      false        Hundreds 2023-02-16
#> 117      PHP  Business    18.723        3      false        Hundreds 2023-02-17
#> 118      PHP  Business    18.693        3      false        Hundreds 2023-02-20
#> 119      PHP  Business    18.654        3      false        Hundreds 2023-02-21
#> 120      PHP  Business    18.699        3      false        Hundreds 2023-02-22
#> 121      PKR  Business     3.873        3      false        Hundreds 2023-02-16
#> 122      PKR  Business     3.946        3      false        Hundreds 2023-02-17
#> 123      PKR  Business     3.926        3      false        Hundreds 2023-02-20
#> 124      PKR  Business     3.914        3      false        Hundreds 2023-02-21
#> 125      PKR  Business     3.938        3      false        Hundreds 2023-02-22
#> 126      PLN  Business    2.2927        4      false           Units 2023-02-16
#> 127      PLN  Business    2.3057        4      false           Units 2023-02-17
#> 128      PLN  Business    2.3123        4      false           Units 2023-02-20
#> 129      PLN  Business    2.3057        4      false           Units 2023-02-21
#> 130      PLN  Business    2.3115        4      false           Units 2023-02-22
#> 131      RON  Business    223.45        2      false        Hundreds 2023-02-16
#> 132      RON  Business    224.04        2      false        Hundreds 2023-02-17
#> 133      RON  Business    223.02        2      false        Hundreds 2023-02-20
#> 134      RON  Business    222.55        2      false        Hundreds 2023-02-21
#> 135      RON  Business    222.72        2      false        Hundreds 2023-02-22
#> 136      SEK  Business     98.29        2      false        Hundreds 2023-02-16
#> 137      SEK  Business     98.36        2      false        Hundreds 2023-02-17
#> 138      SEK  Business     99.15        2      false        Hundreds 2023-02-20
#> 139      SEK  Business     99.43        2      false        Hundreds 2023-02-21
#> 140      SEK  Business     99.55        2      false        Hundreds 2023-02-22
#> 141      SGD  Business    7.6652        4      false           Units 2023-02-16
#> 142      SGD  Business    7.7401        4      false           Units 2023-02-17
#> 143      SGD  Business    7.6922        4      false           Units 2023-02-20
#> 144      SGD  Business    7.6701        4      false           Units 2023-02-21
#> 145      SGD  Business    7.6947        4      false           Units 2023-02-22
#> 146      THB  Business    29.809        3      false        Hundreds 2023-02-16
#> 147      THB  Business     29.98        3      false        Hundreds 2023-02-17
#> 148      THB  Business     29.88        3      false        Hundreds 2023-02-20
#> 149      THB  Business    29.669        3      false        Hundreds 2023-02-21
#> 150      THB  Business    29.821        3      false        Hundreds 2023-02-22
#> 151      TRY  Business      54.3        2      false        Hundreds 2023-02-16
#> 152      TRY  Business     54.94        2      false        Hundreds 2023-02-17
#> 153      TRY  Business     54.47        2      false        Hundreds 2023-02-20
#> 154      TRY  Business      54.4        2      false        Hundreds 2023-02-21
#> 155      TRY  Business     54.58        2      false        Hundreds 2023-02-22
#> 156      TWD  Business    33.758        3      false        Hundreds 2023-02-16
#> 157      TWD  Business    34.052        3      false        Hundreds 2023-02-17
#> 158      TWD  Business    33.796        3      false        Hundreds 2023-02-20
#> 159      TWD  Business    33.693        3      false        Hundreds 2023-02-21
#> 160      TWD  Business    33.777        3      false        Hundreds 2023-02-22
#> 161      TWI  Business    127.14        2       true           Units 2023-02-16
#> 162      TWI  Business    127.83        2       true           Units 2023-02-17
#> 163      TWI  Business    127.61        2       true           Units 2023-02-20
#> 164      TWI  Business    127.51        2       true           Units 2023-02-21
#> 165      TWI  Business    127.75        2       true           Units 2023-02-22
#> 166      USD  Business   10.2349        4      false           Units 2023-02-16
#> 167      USD  Business   10.3671        4      false           Units 2023-02-17
#> 168      USD  Business   10.2757        4      false           Units 2023-02-20
#> 169      USD  Business   10.2652        4      false           Units 2023-02-21
#> 170      USD  Business    10.303        4      false           Units 2023-02-22
#> 171      XDR  Business  13.67618        5       true           Units 2023-02-16
#> 172      XDR  Business  13.80472        5       true           Units 2023-02-17
#> 173      XDR  Business  13.71865        5       true           Units 2023-02-20
#> 174      XDR  Business  13.69769        5       true           Units 2023-02-21
#> 175      XDR  Business  13.73712        5       true           Units 2023-02-22
#> 176      ZAR  Business    0.5643        4      false           Units 2023-02-16
#> 177      ZAR  Business    0.5718        4      false           Units 2023-02-17
#> 178      ZAR  Business    0.5669        4      false           Units 2023-02-20
#> 179      ZAR  Business    0.5626        4      false           Units 2023-02-21
#> 180      ZAR  Business    0.5651        4      false           Units 2023-02-22
#> 181      BYN  Business    4.0541        4      false           Units 2023-02-16
#> 182      BYN  Business    4.1072        4      false           Units 2023-02-17
#> 183      BYN  Business     4.071        4      false           Units 2023-02-20
#> 184      BYN  Business    4.0669        4      false           Units 2023-02-21
#> 185      BYN  Business    4.0818        4      false           Units 2023-02-22
#> 186      BDT  Business      9.67        2      false        Hundreds 2023-02-16
#> 187      BDT  Business       9.8        2      false        Hundreds 2023-02-17
#> 188      BDT  Business      9.68        2      false        Hundreds 2023-02-20
#> 189      BDT  Business      9.67        2      false        Hundreds 2023-02-21
#> 190      BDT  Business      9.62        2      false        Hundreds 2023-02-22
#> 191      MMK  Business    0.4874        4      false        Hundreds 2023-02-16
#> 192      MMK  Business    0.4937        4      false        Hundreds 2023-02-17
#> 193      MMK  Business    0.4893        4      false        Hundreds 2023-02-20
#> 194      MMK  Business    0.4888        4      false        Hundreds 2023-02-21
#> 195      MMK  Business    0.4906        4      false        Hundreds 2023-02-22
#> 196      ISK  Business       7.1        2      false        Hundreds 2023-02-16
#> 197      ISK  Business      7.15        2      false        Hundreds 2023-02-17
#> 198      ISK  Business      7.09        2      false        Hundreds 2023-02-20
#> 199      ISK  Business       7.1        2      false        Hundreds 2023-02-21
#> 200      ISK  Business      7.11        2      false        Hundreds 2023-02-22
#> 201      VND  Business    0.0432        4      false        Hundreds 2023-02-16
#> 202      VND  Business    0.0435        4      false        Hundreds 2023-02-17
#> 203      VND  Business    0.0433        4      false        Hundreds 2023-02-20
#> 204      VND  Business    0.0433        4      false        Hundreds 2023-02-21
#> 205      VND  Business    0.0433        4      false        Hundreds 2023-02-22
#>        date_to       date
#> 1   2022-02-23 2022-02-23
#> 2   2022-02-24 2022-02-24
#> 3   2022-02-25 2022-02-25
#> 4   2022-02-28 2022-02-28
#> 5   2022-03-01 2022-03-01
#> 6   2022-12-23 2022-12-23
#> 7   2022-12-27 2022-12-27
#> 8   2022-12-28 2022-12-28
#> 9   2022-12-29 2022-12-29
#> 10  2022-12-30 2022-12-30
#> 11  2023-02-16 2023-02-16
#> 12  2023-02-17 2023-02-17
#> 13  2023-02-20 2023-02-20
#> 14  2023-02-21 2023-02-21
#> 15  2023-02-22 2023-02-22
#> 16  2023-02-16 2023-02-16
#> 17  2023-02-17 2023-02-17
#> 18  2023-02-20 2023-02-20
#> 19  2023-02-21 2023-02-21
#> 20  2023-02-22 2023-02-22
#> 21  2023-02-16 2023-02-16
#> 22  2023-02-17 2023-02-17
#> 23  2023-02-20 2023-02-20
#> 24  2023-02-21 2023-02-21
#> 25  2023-02-22 2023-02-22
#> 26  2023-02-16 2023-02-16
#> 27  2023-02-17 2023-02-17
#> 28  2023-02-20 2023-02-20
#> 29  2023-02-21 2023-02-21
#> 30  2023-02-22 2023-02-22
#> 31  2023-02-16 2023-02-16
#> 32  2023-02-17 2023-02-17
#> 33  2023-02-20 2023-02-20
#> 34  2023-02-21 2023-02-21
#> 35  2023-02-22 2023-02-22
#> 36  2023-02-16 2023-02-16
#> 37  2023-02-17 2023-02-17
#> 38  2023-02-20 2023-02-20
#> 39  2023-02-21 2023-02-21
#> 40  2023-02-22 2023-02-22
#> 41  2023-02-16 2023-02-16
#> 42  2023-02-17 2023-02-17
#> 43  2023-02-20 2023-02-20
#> 44  2023-02-21 2023-02-21
#> 45  2023-02-22 2023-02-22
#> 46  2023-02-16 2023-02-16
#> 47  2023-02-17 2023-02-17
#> 48  2023-02-20 2023-02-20
#> 49  2023-02-21 2023-02-21
#> 50  2023-02-22 2023-02-22
#> 51  2023-02-16 2023-02-16
#> 52  2023-02-17 2023-02-17
#> 53  2023-02-20 2023-02-20
#> 54  2023-02-21 2023-02-21
#> 55  2023-02-22 2023-02-22
#> 56  2023-02-16 2023-02-16
#> 57  2023-02-17 2023-02-17
#> 58  2023-02-20 2023-02-20
#> 59  2023-02-21 2023-02-21
#> 60  2023-02-22 2023-02-22
#> 61  2023-02-16 2023-02-16
#> 62  2023-02-17 2023-02-17
#> 63  2023-02-20 2023-02-20
#> 64  2023-02-21 2023-02-21
#> 65  2023-02-22 2023-02-22
#> 66  2023-02-16 2023-02-16
#> 67  2023-02-17 2023-02-17
#> 68  2023-02-20 2023-02-20
#> 69  2023-02-21 2023-02-21
#> 70  2023-02-22 2023-02-22
#> 71  2023-02-16 2023-02-16
#> 72  2023-02-17 2023-02-17
#> 73  2023-02-20 2023-02-20
#> 74  2023-02-21 2023-02-21
#> 75  2023-02-22 2023-02-22
#> 76  2023-02-16 2023-02-16
#> 77  2023-02-17 2023-02-17
#> 78  2023-02-20 2023-02-20
#> 79  2023-02-21 2023-02-21
#> 80  2023-02-22 2023-02-22
#> 81  2023-02-16 2023-02-16
#> 82  2023-02-17 2023-02-17
#> 83  2023-02-20 2023-02-20
#> 84  2023-02-21 2023-02-21
#> 85  2023-02-22 2023-02-22
#> 86  2023-02-16 2023-02-16
#> 87  2023-02-17 2023-02-17
#> 88  2023-02-20 2023-02-20
#> 89  2023-02-21 2023-02-21
#> 90  2023-02-22 2023-02-22
#> 91  2023-02-16 2023-02-16
#> 92  2023-02-17 2023-02-17
#> 93  2023-02-20 2023-02-20
#> 94  2023-02-21 2023-02-21
#> 95  2023-02-22 2023-02-22
#> 96  2023-02-16 2023-02-16
#> 97  2023-02-17 2023-02-17
#> 98  2023-02-20 2023-02-20
#> 99  2023-02-21 2023-02-21
#> 100 2023-02-22 2023-02-22
#> 101 2023-02-16 2023-02-16
#> 102 2023-02-17 2023-02-17
#> 103 2023-02-20 2023-02-20
#> 104 2023-02-21 2023-02-21
#> 105 2023-02-22 2023-02-22
#> 106 2023-02-16 2023-02-16
#> 107 2023-02-17 2023-02-17
#> 108 2023-02-20 2023-02-20
#> 109 2023-02-21 2023-02-21
#> 110 2023-02-22 2023-02-22
#> 111 2023-02-16 2023-02-16
#> 112 2023-02-17 2023-02-17
#> 113 2023-02-20 2023-02-20
#> 114 2023-02-21 2023-02-21
#> 115 2023-02-22 2023-02-22
#> 116 2023-02-16 2023-02-16
#> 117 2023-02-17 2023-02-17
#> 118 2023-02-20 2023-02-20
#> 119 2023-02-21 2023-02-21
#> 120 2023-02-22 2023-02-22
#> 121 2023-02-16 2023-02-16
#> 122 2023-02-17 2023-02-17
#> 123 2023-02-20 2023-02-20
#> 124 2023-02-21 2023-02-21
#> 125 2023-02-22 2023-02-22
#> 126 2023-02-16 2023-02-16
#> 127 2023-02-17 2023-02-17
#> 128 2023-02-20 2023-02-20
#> 129 2023-02-21 2023-02-21
#> 130 2023-02-22 2023-02-22
#> 131 2023-02-16 2023-02-16
#> 132 2023-02-17 2023-02-17
#> 133 2023-02-20 2023-02-20
#> 134 2023-02-21 2023-02-21
#> 135 2023-02-22 2023-02-22
#> 136 2023-02-16 2023-02-16
#> 137 2023-02-17 2023-02-17
#> 138 2023-02-20 2023-02-20
#> 139 2023-02-21 2023-02-21
#> 140 2023-02-22 2023-02-22
#> 141 2023-02-16 2023-02-16
#> 142 2023-02-17 2023-02-17
#> 143 2023-02-20 2023-02-20
#> 144 2023-02-21 2023-02-21
#> 145 2023-02-22 2023-02-22
#> 146 2023-02-16 2023-02-16
#> 147 2023-02-17 2023-02-17
#> 148 2023-02-20 2023-02-20
#> 149 2023-02-21 2023-02-21
#> 150 2023-02-22 2023-02-22
#> 151 2023-02-16 2023-02-16
#> 152 2023-02-17 2023-02-17
#> 153 2023-02-20 2023-02-20
#> 154 2023-02-21 2023-02-21
#> 155 2023-02-22 2023-02-22
#> 156 2023-02-16 2023-02-16
#> 157 2023-02-17 2023-02-17
#> 158 2023-02-20 2023-02-20
#> 159 2023-02-21 2023-02-21
#> 160 2023-02-22 2023-02-22
#> 161 2023-02-16 2023-02-16
#> 162 2023-02-17 2023-02-17
#> 163 2023-02-20 2023-02-20
#> 164 2023-02-21 2023-02-21
#> 165 2023-02-22 2023-02-22
#> 166 2023-02-16 2023-02-16
#> 167 2023-02-17 2023-02-17
#> 168 2023-02-20 2023-02-20
#> 169 2023-02-21 2023-02-21
#> 170 2023-02-22 2023-02-22
#> 171 2023-02-16 2023-02-16
#> 172 2023-02-17 2023-02-17
#> 173 2023-02-20 2023-02-20
#> 174 2023-02-21 2023-02-21
#> 175 2023-02-22 2023-02-22
#> 176 2023-02-16 2023-02-16
#> 177 2023-02-17 2023-02-17
#> 178 2023-02-20 2023-02-20
#> 179 2023-02-21 2023-02-21
#> 180 2023-02-22 2023-02-22
#> 181 2023-02-16 2023-02-16
#> 182 2023-02-17 2023-02-17
#> 183 2023-02-20 2023-02-20
#> 184 2023-02-21 2023-02-21
#> 185 2023-02-22 2023-02-22
#> 186 2023-02-16 2023-02-16
#> 187 2023-02-17 2023-02-17
#> 188 2023-02-20 2023-02-20
#> 189 2023-02-21 2023-02-21
#> 190 2023-02-22 2023-02-22
#> 191 2023-02-16 2023-02-16
#> 192 2023-02-17 2023-02-17
#> 193 2023-02-20 2023-02-20
#> 194 2023-02-21 2023-02-21
#> 195 2023-02-22 2023-02-22
#> 196 2023-02-16 2023-02-16
#> 197 2023-02-17 2023-02-17
#> 198 2023-02-20 2023-02-20
#> 199 2023-02-21 2023-02-21
#> 200 2023-02-22 2023-02-22
#> 201 2023-02-16 2023-02-16
#> 202 2023-02-17 2023-02-17
#> 203 2023-02-20 2023-02-20
#> 204 2023-02-21 2023-02-21
#> 205 2023-02-22 2023-02-22