Get exchange rate

Fetch exchange rates from Norges Bank.

Usage

get_exchange_rate(
  currency,
  frequency = c("daily", "monthly", "annual"),
  n_obs = 10,
  start = NULL,
  end = NULL,
  simplify = TRUE,
  raw_response = FALSE
)

Arguments

currency

The currency to fetch data for.

frequency

Frequency of data. Can be daily, monthly or annual. Defaults to daily.

n_obs

The number of observations to fetch. See details.

start

The first date to fetch data for.

end

The last date to fetch data for.

simplify

If TRUE (default), a single data.frame is returned. Ignored if raw_response is set to TRUE.

raw_response

If TRUE a list of class noAPI is returned, including the raw httr2_response.

Value

data.frame or list

Details

Currency must be formatted as a three letter ISO 3166 code. A list of available country codes can be found at Norges Bank. You can fetch observations for several currencies by passing a character vector. If you want to include all currencies, you can set the currency argument to ALL. Note that the number of observations is limited to 20 when retrieving all currencies.

By default, the 10 latest observations are returned. You can adjust the number of observations by setting n_obs to a positive integer. You can also fetch exchange rates for a specific time period by setting the arguments start and end. Both arguments accept a date in the format %Y-%m-%d (YYYY-MM-DD).

The function returns a data.frame by default. If you prefer the output as a list you can set simplify to FALSE. This can be useful to keep programmatically track of failed queries. If you set raw_response to TRUE, the raw response from the API will be returned together with the parsed response. Note that the response will then be returned silently.

Examples

# Get latest daily exchange rates
get_exchange_rate('EUR')
#>    currency frequency exch_rate decimals calculated unit_multiplier  date_from
#> 1       EUR  Business   11.7220        4      false           Units 2023-04-26
#> 2       EUR  Business   11.7140        4      false           Units 2023-04-27
#> 3       EUR  Business   11.7910        4      false           Units 2023-04-28
#> 4       EUR  Business   11.7620        4      false           Units 2023-05-02
#> 5       EUR  Business   11.8860        4      false           Units 2023-05-03
#> 6       EUR  Business   11.8282        4      false           Units 2023-05-04
#> 7       EUR  Business   11.6895        4      false           Units 2023-05-05
#> 8       EUR  Business   11.5505        4      false           Units 2023-05-08
#> 9       EUR  Business   11.5750        4      false           Units 2023-05-09
#> 10      EUR  Business   11.5485        4      false           Units 2023-05-10
#>       date_to       date
#> 1  2023-04-26 2023-04-26
#> 2  2023-04-27 2023-04-27
#> 3  2023-04-28 2023-04-28
#> 4  2023-05-02 2023-05-02
#> 5  2023-05-03 2023-05-03
#> 6  2023-05-04 2023-05-04
#> 7  2023-05-05 2023-05-05
#> 8  2023-05-08 2023-05-08
#> 9  2023-05-09 2023-05-09
#> 10 2023-05-10 2023-05-10
# Get monthly exchange rates for multiple currencies
get_exchange_rate(c('EUR', 'USD'), frequency = 'monthly')
#>    currency frequency exch_rate decimals calculated unit_multiplier  date_from
#> 1       EUR   Monthly   10.1823        4      false           Units 2022-07-01
#> 2       EUR   Monthly    9.8309        4      false           Units 2022-08-01
#> 3       EUR   Monthly   10.1697        4      false           Units 2022-09-01
#> 4       EUR   Monthly   10.3919        4      false           Units 2022-10-01
#> 5       EUR   Monthly   10.3357        4      false           Units 2022-11-01
#> 6       EUR   Monthly   10.4480        4      false           Units 2022-12-01
#> 7       EUR   Monthly   10.7149        4      false           Units 2023-01-01
#> 8       EUR   Monthly   10.9529        4      false           Units 2023-02-01
#> 9       EUR   Monthly   11.2858        4      false           Units 2023-03-01
#> 10      EUR   Monthly   11.5265        4      false           Units 2023-04-01
#> 11      USD   Monthly   10.0042        4      false           Units 2022-07-01
#> 12      USD   Monthly    9.7068        4      false           Units 2022-08-01
#> 13      USD   Monthly   10.2735        4      false           Units 2022-09-01
#> 14      USD   Monthly   10.5775        4      false           Units 2022-10-01
#> 15      USD   Monthly   10.1355        4      false           Units 2022-11-01
#> 16      USD   Monthly    9.8672        4      false           Units 2022-12-01
#> 17      USD   Monthly    9.9506        4      false           Units 2023-01-01
#> 18      USD   Monthly   10.2231        4      false           Units 2023-02-01
#> 19      USD   Monthly   10.5425        4      false           Units 2023-03-01
#> 20      USD   Monthly   10.5064        4      false           Units 2023-04-01
#>       date_to    date
#> 1  2022-07-31 2022-07
#> 2  2022-08-31 2022-08
#> 3  2022-09-30 2022-09
#> 4  2022-10-31 2022-10
#> 5  2022-11-30 2022-11
#> 6  2022-12-31 2022-12
#> 7  2023-01-31 2023-01
#> 8  2023-02-28 2023-02
#> 9  2023-03-31 2023-03
#> 10 2023-04-30 2023-04
#> 11 2022-07-31 2022-07
#> 12 2022-08-31 2022-08
#> 13 2022-09-30 2022-09
#> 14 2022-10-31 2022-10
#> 15 2022-11-30 2022-11
#> 16 2022-12-31 2022-12
#> 17 2023-01-31 2023-01
#> 18 2023-02-28 2023-02
#> 19 2023-03-31 2023-03
#> 20 2023-04-30 2023-04
# Get last 5 observations for all currencies
get_exchange_rate('ALL', n_obs = 5)
#>     currency frequency   exch_rate decimals calculated unit_multiplier
#> 1        HUF  Business    3.163100        4      false        Hundreds
#> 2        HUF  Business    3.138300        4      false        Hundreds
#> 3        HUF  Business    3.103300        4      false        Hundreds
#> 4        HUF  Business    3.114200        4      false        Hundreds
#> 5        HUF  Business    3.113000        4      false        Hundreds
#> 6        DKK  Business  158.760000        2      false        Hundreds
#> 7        DKK  Business  156.910000        2      false        Hundreds
#> 8        DKK  Business  155.100000        2      false        Hundreds
#> 9        DKK  Business  155.470000        2      false        Hundreds
#> 10       DKK  Business  155.060000        2      false        Hundreds
#> 11       RUB  Business   11.040000        3      false        Hundreds
#> 12       RUB  Business   10.539000        3      false        Hundreds
#> 13       RUB  Business   10.777000        3      false        Hundreds
#> 14       RUB  Business    8.613000        3      false        Hundreds
#> 15       RUB  Business    8.413000        3      false        Hundreds
#> 16       MMK  Business    0.508600        4      false        Hundreds
#> 17       MMK  Business    0.505400        4      false        Hundreds
#> 18       MMK  Business    0.498300        4      false        Hundreds
#> 19       MMK  Business    0.503000        4      false        Hundreds
#> 20       MMK  Business    0.502200        4      false        Hundreds
#> 21       XDR  Business   14.450410        5       true           Units
#> 22       XDR  Business   14.339430        5       true           Units
#> 23       XDR  Business   14.143470        5       true           Units
#> 24       XDR  Business   14.238740        5       true           Units
#> 25       XDR  Business   14.212630        5       true           Units
#> 26       TRY  Business   54.800000        2      false        Hundreds
#> 27       TRY  Business   54.430000        2      false        Hundreds
#> 28       TRY  Business   53.650000        2      false        Hundreds
#> 29       TRY  Business   54.120000        2      false        Hundreds
#> 30       TRY  Business   53.970000        2      false        Hundreds
#> 31       RON  Business  239.950000        2      false        Hundreds
#> 32       RON  Business  237.210000        2      false        Hundreds
#> 33       RON  Business  234.550000        2      false        Hundreds
#> 34       RON  Business  235.260000        2      false        Hundreds
#> 35       RON  Business  234.630000        2      false        Hundreds
#> 36       NZD  Business    6.694700        4      false           Units
#> 37       NZD  Business    6.684300        4      false           Units
#> 38       NZD  Business    6.642400        4      false           Units
#> 39       NZD  Business    6.683000        4      false           Units
#> 40       NZD  Business    6.674300        4      false           Units
#> 41       JPY  Business    7.942700        4      false        Hundreds
#> 42       JPY  Business    7.909000        4      false        Hundreds
#> 43       JPY  Business    7.744200        4      false        Hundreds
#> 44       JPY  Business    7.822500        4      false        Hundreds
#> 45       JPY  Business    7.793600        4      false        Hundreds
#> 46       HKD  Business    1.361000        4      false           Units
#> 47       HKD  Business    1.352400        4      false           Units
#> 48       HKD  Business    1.333500        4      false           Units
#> 49       HKD  Business    1.347000        4      false           Units
#> 50       HKD  Business    1.346900        4      false           Units
#> 51       CZK  Business   50.419000        3      false        Hundreds
#> 52       CZK  Business   49.953000        3      false        Hundreds
#> 53       CZK  Business   49.340000        3      false        Hundreds
#> 54       CZK  Business   49.544000        3      false        Hundreds
#> 55       CZK  Business   49.313000        3      false        Hundreds
#> 56       CAD  Business    7.847800        4      false           Units
#> 57       CAD  Business    7.866400        4      false           Units
#> 58       CAD  Business    7.852100        4      false           Units
#> 59       CAD  Business    7.885400        4      false           Units
#> 60       CAD  Business    7.879200        4      false           Units
#> 61       BDT  Business   10.030000        2      false        Hundreds
#> 62       BDT  Business    9.970000        2      false        Hundreds
#> 63       BDT  Business    9.780000        2      false        Hundreds
#> 64       BDT  Business    9.760000        2      false        Hundreds
#> 65       BDT  Business    9.820000        2      false        Hundreds
#> 66       USD  Business   10.681100        4      false           Units
#> 67       USD  Business   10.613300        4      false           Units
#> 68       USD  Business   10.465300        4      false           Units
#> 69       USD  Business   10.562100        4      false           Units
#> 70       USD  Business   10.546600        4      false           Units
#> 71       THB  Business   31.601000        3      false        Hundreds
#> 72       THB  Business   31.447000        3      false        Hundreds
#> 73       THB  Business   30.899000        3      false        Hundreds
#> 74       THB  Business   31.351000        3      false        Hundreds
#> 75       THB  Business   31.332000        3      false        Hundreds
#> 76       PLN  Business    2.576700        4      false           Units
#> 77       PLN  Business    2.554000        4      false           Units
#> 78       PLN  Business    2.527800        4      false           Units
#> 79       PLN  Business    2.531500        4      false           Units
#> 80       PLN  Business    2.545800        4      false           Units
#> 81       AUD  Business    7.131900        4      false           Units
#> 82       AUD  Business    7.136900        4      false           Units
#> 83       AUD  Business    7.110600        4      false           Units
#> 84       AUD  Business    7.137600        4      false           Units
#> 85       AUD  Business    7.119000        4      false           Units
#> 86       MYR  Business    2.398600        4      false           Units
#> 87       MYR  Business    2.392000        4      false           Units
#> 88       MYR  Business    2.358600        4      false           Units
#> 89       MYR  Business    2.373800        4      false           Units
#> 90       MYR  Business    2.365500        4      false           Units
#> 91       INR  Business   13.063000        3      false        Hundreds
#> 92       INR  Business   12.988000        3      false        Hundreds
#> 93       INR  Business   12.794000        3      false        Hundreds
#> 94       INR  Business   12.868000        3      false        Hundreds
#> 95       INR  Business   12.858000        3      false        Hundreds
#> 96       IDR  Business    0.072728        6      false        Hundreds
#> 97       IDR  Business    0.072352        6      false        Hundreds
#> 98       IDR  Business    0.071135        6      false        Hundreds
#> 99       IDR  Business    0.071631        6      false        Hundreds
#> 100      IDR  Business    0.071503        6      false        Hundreds
#> 101      GBP  Business   13.438800        4      false           Units
#> 102      GBP  Business   13.378100        4      false           Units
#> 103      GBP  Business   13.241700        4      false           Units
#> 104      GBP  Business   13.306100        4      false           Units
#> 105      GBP  Business   13.302700        4      false           Units
#> 106      CNY  Business  154.540000        2      false        Hundreds
#> 107      CNY  Business  153.580000        2      false        Hundreds
#> 108      CNY  Business  151.290000        2      false        Hundreds
#> 109      CNY  Business  152.600000        2      false        Hundreds
#> 110      CNY  Business  152.320000        2      false        Hundreds
#> 111      BRL  Business    2.143000        4      false           Units
#> 112      BRL  Business    2.126900        4      false           Units
#> 113      BRL  Business    2.115000        4      false           Units
#> 114      BRL  Business    2.109800        4      false           Units
#> 115      BRL  Business    2.119100        4      false           Units
#> 116      VND  Business    0.045500        4      false        Hundreds
#> 117      VND  Business    0.045300        4      false        Hundreds
#> 118      VND  Business    0.044600        4      false        Hundreds
#> 119      VND  Business    0.045000        4      false        Hundreds
#> 120      VND  Business    0.044900        4      false        Hundreds
#> 121      BYN  Business    4.231600        4      false           Units
#> 122      BYN  Business    4.204800        4      false           Units
#> 123      BYN  Business    4.146100        4      false           Units
#> 124      BYN  Business    4.184500        4      false           Units
#> 125      BYN  Business    4.178400        4      false           Units
#> 126      TWI  Business  135.650000        2       true           Units
#> 127      TWI  Business  134.650000        2       true           Units
#> 128      TWI  Business  133.110000        2       true           Units
#> 129      TWI  Business  133.610000        2       true           Units
#> 130      TWI  Business  133.400000        2       true           Units
#> 131      SGD  Business    8.049100        4      false           Units
#> 132      SGD  Business    8.006500        4      false           Units
#> 133      SGD  Business    7.900500        4      false           Units
#> 134      SGD  Business    7.958600        4      false           Units
#> 135      SGD  Business    7.942000        4      false           Units
#> 136      PKR  Business    3.764000        3      false        Hundreds
#> 137      PKR  Business    3.745000        3      false        Hundreds
#> 138      PKR  Business    3.687000        3      false        Hundreds
#> 139      PKR  Business    3.701000        3      false        Hundreds
#> 140      PKR  Business    3.638000        3      false        Hundreds
#> 141      MXN  Business   59.680000        2      false        Hundreds
#> 142      MXN  Business   59.320000        2      false        Hundreds
#> 143      MXN  Business   58.810000        2      false        Hundreds
#> 144      MXN  Business   59.370000        2      false        Hundreds
#> 145      MXN  Business   59.540000        2      false        Hundreds
#> 146      ILS  Business    2.937400        4      false           Units
#> 147      ILS  Business    2.910800        4      false           Units
#> 148      ILS  Business    2.878100        4      false           Units
#> 149      ILS  Business    2.888200        4      false           Units
#> 150      ILS  Business    2.876100        4      false           Units
#> 151      I44  Business  123.790000        2       true           Units
#> 152      I44  Business  122.880000        2       true           Units
#> 153      I44  Business  121.420000        2       true           Units
#> 154      I44  Business  121.960000        2       true           Units
#> 155      I44  Business  121.820000        2       true           Units
#> 156      EUR  Business   11.828200        4      false           Units
#> 157      EUR  Business   11.689500        4      false           Units
#> 158      EUR  Business   11.550500        4      false           Units
#> 159      EUR  Business   11.575000        4      false           Units
#> 160      EUR  Business   11.548500        4      false           Units
#> 161      CHF  Business 1206.710000        2      false        Hundreds
#> 162      CHF  Business 1189.410000        2      false        Hundreds
#> 163      CHF  Business 1177.060000        2      false        Hundreds
#> 164      CHF  Business 1182.570000        2      false        Hundreds
#> 165      CHF  Business 1181.920000        2      false        Hundreds
#> 166      HRK  Business  138.580000        2      false        Hundreds
#> 167      HRK  Business  139.160000        2      false        Hundreds
#> 168      HRK  Business  138.650000        2      false        Hundreds
#> 169      HRK  Business  139.990000        2      false        Hundreds
#> 170      HRK  Business  139.510000        2      false        Hundreds
#> 171      BGN  Business  604.780000        2      false        Hundreds
#> 172      BGN  Business  597.680000        2      false        Hundreds
#> 173      BGN  Business  590.580000        2      false        Hundreds
#> 174      BGN  Business  591.830000        2      false        Hundreds
#> 175      BGN  Business  590.470000        2      false        Hundreds
#> 176      ISK  Business    7.880000        2      false        Hundreds
#> 177      ISK  Business    7.790000        2      false        Hundreds
#> 178      ISK  Business    7.670000        2      false        Hundreds
#> 179      ISK  Business    7.680000        2      false        Hundreds
#> 180      ISK  Business    7.700000        2      false        Hundreds
#> 181      ZAR  Business    0.587400        4      false           Units
#> 182      ZAR  Business    0.577700        4      false           Units
#> 183      ZAR  Business    0.570200        4      false           Units
#> 184      ZAR  Business    0.573000        4      false           Units
#> 185      ZAR  Business    0.563300        4      false           Units
#> 186      TWD  Business   34.809000        3      false        Hundreds
#> 187      TWD  Business   34.606000        3      false        Hundreds
#> 188      TWD  Business   34.102000        3      false        Hundreds
#> 189      TWD  Business   34.385000        3      false        Hundreds
#> 190      TWD  Business   34.363000        3      false        Hundreds
#> 191      SEK  Business  104.300000        2      false        Hundreds
#> 192      SEK  Business  104.030000        2      false        Hundreds
#> 193      SEK  Business  103.120000        2      false        Hundreds
#> 194      SEK  Business  103.450000        2      false        Hundreds
#> 195      SEK  Business  103.170000        2      false        Hundreds
#> 196      PHP  Business   19.300000        3      false        Hundreds
#> 197      PHP  Business   19.217000        3      false        Hundreds
#> 198      PHP  Business   18.931000        3      false        Hundreds
#> 199      PHP  Business   18.937000        3      false        Hundreds
#> 200      PHP  Business   18.934000        3      false        Hundreds
#> 201      KRW  Business    0.807000        4      false        Hundreds
#> 202      KRW  Business    0.805100        4      false        Hundreds
#> 203      KRW  Business    0.792100        4      false        Hundreds
#> 204      KRW  Business    0.797200        4      false        Hundreds
#> 205      KRW  Business    0.796600        4      false        Hundreds
#>      date_from    date_to       date
#> 1   2023-05-04 2023-05-04 2023-05-04
#> 2   2023-05-05 2023-05-05 2023-05-05
#> 3   2023-05-08 2023-05-08 2023-05-08
#> 4   2023-05-09 2023-05-09 2023-05-09
#> 5   2023-05-10 2023-05-10 2023-05-10
#> 6   2023-05-04 2023-05-04 2023-05-04
#> 7   2023-05-05 2023-05-05 2023-05-05
#> 8   2023-05-08 2023-05-08 2023-05-08
#> 9   2023-05-09 2023-05-09 2023-05-09
#> 10  2023-05-10 2023-05-10 2023-05-10
#> 11  2022-02-23 2022-02-23 2022-02-23
#> 12  2022-02-24 2022-02-24 2022-02-24
#> 13  2022-02-25 2022-02-25 2022-02-25
#> 14  2022-02-28 2022-02-28 2022-02-28
#> 15  2022-03-01 2022-03-01 2022-03-01
#> 16  2023-05-04 2023-05-04 2023-05-04
#> 17  2023-05-05 2023-05-05 2023-05-05
#> 18  2023-05-08 2023-05-08 2023-05-08
#> 19  2023-05-09 2023-05-09 2023-05-09
#> 20  2023-05-10 2023-05-10 2023-05-10
#> 21  2023-05-04 2023-05-04 2023-05-04
#> 22  2023-05-05 2023-05-05 2023-05-05
#> 23  2023-05-08 2023-05-08 2023-05-08
#> 24  2023-05-09 2023-05-09 2023-05-09
#> 25  2023-05-10 2023-05-10 2023-05-10
#> 26  2023-05-04 2023-05-04 2023-05-04
#> 27  2023-05-05 2023-05-05 2023-05-05
#> 28  2023-05-08 2023-05-08 2023-05-08
#> 29  2023-05-09 2023-05-09 2023-05-09
#> 30  2023-05-10 2023-05-10 2023-05-10
#> 31  2023-05-04 2023-05-04 2023-05-04
#> 32  2023-05-05 2023-05-05 2023-05-05
#> 33  2023-05-08 2023-05-08 2023-05-08
#> 34  2023-05-09 2023-05-09 2023-05-09
#> 35  2023-05-10 2023-05-10 2023-05-10
#> 36  2023-05-04 2023-05-04 2023-05-04
#> 37  2023-05-05 2023-05-05 2023-05-05
#> 38  2023-05-08 2023-05-08 2023-05-08
#> 39  2023-05-09 2023-05-09 2023-05-09
#> 40  2023-05-10 2023-05-10 2023-05-10
#> 41  2023-05-04 2023-05-04 2023-05-04
#> 42  2023-05-05 2023-05-05 2023-05-05
#> 43  2023-05-08 2023-05-08 2023-05-08
#> 44  2023-05-09 2023-05-09 2023-05-09
#> 45  2023-05-10 2023-05-10 2023-05-10
#> 46  2023-05-04 2023-05-04 2023-05-04
#> 47  2023-05-05 2023-05-05 2023-05-05
#> 48  2023-05-08 2023-05-08 2023-05-08
#> 49  2023-05-09 2023-05-09 2023-05-09
#> 50  2023-05-10 2023-05-10 2023-05-10
#> 51  2023-05-04 2023-05-04 2023-05-04
#> 52  2023-05-05 2023-05-05 2023-05-05
#> 53  2023-05-08 2023-05-08 2023-05-08
#> 54  2023-05-09 2023-05-09 2023-05-09
#> 55  2023-05-10 2023-05-10 2023-05-10
#> 56  2023-05-04 2023-05-04 2023-05-04
#> 57  2023-05-05 2023-05-05 2023-05-05
#> 58  2023-05-08 2023-05-08 2023-05-08
#> 59  2023-05-09 2023-05-09 2023-05-09
#> 60  2023-05-10 2023-05-10 2023-05-10
#> 61  2023-05-04 2023-05-04 2023-05-04
#> 62  2023-05-05 2023-05-05 2023-05-05
#> 63  2023-05-08 2023-05-08 2023-05-08
#> 64  2023-05-09 2023-05-09 2023-05-09
#> 65  2023-05-10 2023-05-10 2023-05-10
#> 66  2023-05-04 2023-05-04 2023-05-04
#> 67  2023-05-05 2023-05-05 2023-05-05
#> 68  2023-05-08 2023-05-08 2023-05-08
#> 69  2023-05-09 2023-05-09 2023-05-09
#> 70  2023-05-10 2023-05-10 2023-05-10
#> 71  2023-05-04 2023-05-04 2023-05-04
#> 72  2023-05-05 2023-05-05 2023-05-05
#> 73  2023-05-08 2023-05-08 2023-05-08
#> 74  2023-05-09 2023-05-09 2023-05-09
#> 75  2023-05-10 2023-05-10 2023-05-10
#> 76  2023-05-04 2023-05-04 2023-05-04
#> 77  2023-05-05 2023-05-05 2023-05-05
#> 78  2023-05-08 2023-05-08 2023-05-08
#> 79  2023-05-09 2023-05-09 2023-05-09
#> 80  2023-05-10 2023-05-10 2023-05-10
#> 81  2023-05-04 2023-05-04 2023-05-04
#> 82  2023-05-05 2023-05-05 2023-05-05
#> 83  2023-05-08 2023-05-08 2023-05-08
#> 84  2023-05-09 2023-05-09 2023-05-09
#> 85  2023-05-10 2023-05-10 2023-05-10
#> 86  2023-05-04 2023-05-04 2023-05-04
#> 87  2023-05-05 2023-05-05 2023-05-05
#> 88  2023-05-08 2023-05-08 2023-05-08
#> 89  2023-05-09 2023-05-09 2023-05-09
#> 90  2023-05-10 2023-05-10 2023-05-10
#> 91  2023-05-04 2023-05-04 2023-05-04
#> 92  2023-05-05 2023-05-05 2023-05-05
#> 93  2023-05-08 2023-05-08 2023-05-08
#> 94  2023-05-09 2023-05-09 2023-05-09
#> 95  2023-05-10 2023-05-10 2023-05-10
#> 96  2023-05-04 2023-05-04 2023-05-04
#> 97  2023-05-05 2023-05-05 2023-05-05
#> 98  2023-05-08 2023-05-08 2023-05-08
#> 99  2023-05-09 2023-05-09 2023-05-09
#> 100 2023-05-10 2023-05-10 2023-05-10
#> 101 2023-05-04 2023-05-04 2023-05-04
#> 102 2023-05-05 2023-05-05 2023-05-05
#> 103 2023-05-08 2023-05-08 2023-05-08
#> 104 2023-05-09 2023-05-09 2023-05-09
#> 105 2023-05-10 2023-05-10 2023-05-10
#> 106 2023-05-04 2023-05-04 2023-05-04
#> 107 2023-05-05 2023-05-05 2023-05-05
#> 108 2023-05-08 2023-05-08 2023-05-08
#> 109 2023-05-09 2023-05-09 2023-05-09
#> 110 2023-05-10 2023-05-10 2023-05-10
#> 111 2023-05-04 2023-05-04 2023-05-04
#> 112 2023-05-05 2023-05-05 2023-05-05
#> 113 2023-05-08 2023-05-08 2023-05-08
#> 114 2023-05-09 2023-05-09 2023-05-09
#> 115 2023-05-10 2023-05-10 2023-05-10
#> 116 2023-05-04 2023-05-04 2023-05-04
#> 117 2023-05-05 2023-05-05 2023-05-05
#> 118 2023-05-08 2023-05-08 2023-05-08
#> 119 2023-05-09 2023-05-09 2023-05-09
#> 120 2023-05-10 2023-05-10 2023-05-10
#> 121 2023-05-04 2023-05-04 2023-05-04
#> 122 2023-05-05 2023-05-05 2023-05-05
#> 123 2023-05-08 2023-05-08 2023-05-08
#> 124 2023-05-09 2023-05-09 2023-05-09
#> 125 2023-05-10 2023-05-10 2023-05-10
#> 126 2023-05-04 2023-05-04 2023-05-04
#> 127 2023-05-05 2023-05-05 2023-05-05
#> 128 2023-05-08 2023-05-08 2023-05-08
#> 129 2023-05-09 2023-05-09 2023-05-09
#> 130 2023-05-10 2023-05-10 2023-05-10
#> 131 2023-05-04 2023-05-04 2023-05-04
#> 132 2023-05-05 2023-05-05 2023-05-05
#> 133 2023-05-08 2023-05-08 2023-05-08
#> 134 2023-05-09 2023-05-09 2023-05-09
#> 135 2023-05-10 2023-05-10 2023-05-10
#> 136 2023-05-04 2023-05-04 2023-05-04
#> 137 2023-05-05 2023-05-05 2023-05-05
#> 138 2023-05-08 2023-05-08 2023-05-08
#> 139 2023-05-09 2023-05-09 2023-05-09
#> 140 2023-05-10 2023-05-10 2023-05-10
#> 141 2023-05-04 2023-05-04 2023-05-04
#> 142 2023-05-05 2023-05-05 2023-05-05
#> 143 2023-05-08 2023-05-08 2023-05-08
#> 144 2023-05-09 2023-05-09 2023-05-09
#> 145 2023-05-10 2023-05-10 2023-05-10
#> 146 2023-05-04 2023-05-04 2023-05-04
#> 147 2023-05-05 2023-05-05 2023-05-05
#> 148 2023-05-08 2023-05-08 2023-05-08
#> 149 2023-05-09 2023-05-09 2023-05-09
#> 150 2023-05-10 2023-05-10 2023-05-10
#> 151 2023-05-04 2023-05-04 2023-05-04
#> 152 2023-05-05 2023-05-05 2023-05-05
#> 153 2023-05-08 2023-05-08 2023-05-08
#> 154 2023-05-09 2023-05-09 2023-05-09
#> 155 2023-05-10 2023-05-10 2023-05-10
#> 156 2023-05-04 2023-05-04 2023-05-04
#> 157 2023-05-05 2023-05-05 2023-05-05
#> 158 2023-05-08 2023-05-08 2023-05-08
#> 159 2023-05-09 2023-05-09 2023-05-09
#> 160 2023-05-10 2023-05-10 2023-05-10
#> 161 2023-05-04 2023-05-04 2023-05-04
#> 162 2023-05-05 2023-05-05 2023-05-05
#> 163 2023-05-08 2023-05-08 2023-05-08
#> 164 2023-05-09 2023-05-09 2023-05-09
#> 165 2023-05-10 2023-05-10 2023-05-10
#> 166 2022-12-23 2022-12-23 2022-12-23
#> 167 2022-12-27 2022-12-27 2022-12-27
#> 168 2022-12-28 2022-12-28 2022-12-28
#> 169 2022-12-29 2022-12-29 2022-12-29
#> 170 2022-12-30 2022-12-30 2022-12-30
#> 171 2023-05-04 2023-05-04 2023-05-04
#> 172 2023-05-05 2023-05-05 2023-05-05
#> 173 2023-05-08 2023-05-08 2023-05-08
#> 174 2023-05-09 2023-05-09 2023-05-09
#> 175 2023-05-10 2023-05-10 2023-05-10
#> 176 2023-05-04 2023-05-04 2023-05-04
#> 177 2023-05-05 2023-05-05 2023-05-05
#> 178 2023-05-08 2023-05-08 2023-05-08
#> 179 2023-05-09 2023-05-09 2023-05-09
#> 180 2023-05-10 2023-05-10 2023-05-10
#> 181 2023-05-04 2023-05-04 2023-05-04
#> 182 2023-05-05 2023-05-05 2023-05-05
#> 183 2023-05-08 2023-05-08 2023-05-08
#> 184 2023-05-09 2023-05-09 2023-05-09
#> 185 2023-05-10 2023-05-10 2023-05-10
#> 186 2023-05-04 2023-05-04 2023-05-04
#> 187 2023-05-05 2023-05-05 2023-05-05
#> 188 2023-05-08 2023-05-08 2023-05-08
#> 189 2023-05-09 2023-05-09 2023-05-09
#> 190 2023-05-10 2023-05-10 2023-05-10
#> 191 2023-05-04 2023-05-04 2023-05-04
#> 192 2023-05-05 2023-05-05 2023-05-05
#> 193 2023-05-08 2023-05-08 2023-05-08
#> 194 2023-05-09 2023-05-09 2023-05-09
#> 195 2023-05-10 2023-05-10 2023-05-10
#> 196 2023-05-04 2023-05-04 2023-05-04
#> 197 2023-05-05 2023-05-05 2023-05-05
#> 198 2023-05-08 2023-05-08 2023-05-08
#> 199 2023-05-09 2023-05-09 2023-05-09
#> 200 2023-05-10 2023-05-10 2023-05-10
#> 201 2023-05-04 2023-05-04 2023-05-04
#> 202 2023-05-05 2023-05-05 2023-05-05
#> 203 2023-05-08 2023-05-08 2023-05-08
#> 204 2023-05-09 2023-05-09 2023-05-09
#> 205 2023-05-10 2023-05-10 2023-05-10